% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{activation_elu_plus_one_plus_epsilon}
\alias{activation_elu_plus_one_plus_epsilon}
\title{Returns a custom activation functions to prevent NaNs in loss function.}
\usage{
activation_elu_plus_one_plus_epsilon(x)
}
\arguments{
\item{x}{input tensor}
}
\value{
an activation function
}
\description{
Ported from:
}
\details{
https://github.com/cpmpercussion/keras-mdn-layer/
}
\examples{

library( keras )

K <- keras::backend()
x <- K$ones( list( 1, 1 ) )
activation <- activation_elu_plus_one_plus_epsilon( x )

}
\author{
Tustison NJ
}
