% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histology.R
\name{allenHistologyBrainMask}
\alias{allenHistologyBrainMask}
\title{Determine brain foreground of Allen's mouse embroyonic data.}
\usage{
allenHistologyBrainMask(
  image,
  whichAxis = 3,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input image}

\item{whichAxis}{If 3-D image, which_axis specifies the direction of the slice.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Determine brain foreground of Allen's mouse embroyonic data.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "image.nii.gz" )
output <- allenHistologyBrainMask( image )
}
}
\author{
Tustison NJ
}
