% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{data}
\name{AnchorBoxLayer2D}
\alias{AnchorBoxLayer2D}
\title{Anchor box layer for SSD architecture (2-D).}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnchorBoxLayer2D
}
\arguments{
\item{imageSize}{size of the input image.}

\item{scale}{scale of each box (in pixels).}

\item{nextScale}{next scale of each box (in pixels).}

\item{aspectRatios}{vector describing the geometries of the anchor
boxes for this layer.}

\item{variances}{explained here:
https://github.com/rykov8/ssd_keras/issues/53

Input shape:}
}
\value{
a 5-D tensor with shape
\link{batchSize, widthSize, heightSize, numberOfBoxes, 8}
In the last dimension, the first 4 values correspond to the
2-D coordinates of the bounding boxes and the other 4 are the variances.
}
\description{
Anchor (or prior) boxes for SSD architecture (2-D).  The input shape is
defined as
}
\details{
Theano:  \link{batchSize, channelSize, widthSize, heightSize}
tensorflow:  \link{batchSize, widthSize, heightSize, channelSize}
}
\examples{

\dontrun{ 

library( keras )

}
}
\author{
Tustison NJ
}
\keyword{datasets}
