% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDBPN4x.R
\name{applyDBPNsr}
\alias{applyDBPNsr}
\title{applyDBPNsr}
\usage{
applyDBPNsr(image, modelFilename, expansionFactor = 4,
  strideLength = 20, lowResolutionPatchSize = c(48, 48),
  standardDeviationOffset = 0.01, verbose = FALSE)
}
\arguments{
\item{image}{input image}

\item{modelFilename}{see \code{getPretrainedNetwork}}

\item{expansionFactor}{size to upscale, should match network pretraining}

\item{strideLength}{stride length should be less than patch size}

\item{lowResolutionPatchSize}{patch size at low resolution that will be
expanded by expansionFactor - needs to be consistent with trained model.}

\item{standardDeviationOffset}{an offset added to the normalization of patches to prevent divisions by zero}

\item{verbose}{If \code{TRUE}, show status messages}
}
\value{
image upscaled to resolution provided by network
}
\description{
Apply pretrained DBPN super-resolution network
}
\examples{
\dontrun{
simg <- applyDBPNsr( ri( 1 ), getPretrainedNetwork( "dbpn4x" ) )
}
}
\author{
Avants BB
}
