% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDBPN4x.R
\name{applySuperResolutionModel}
\alias{applySuperResolutionModel}
\title{applySuperResolutionModel}
\usage{
applySuperResolutionModel(image, model, strideLength = 20,
  targetRange = c(-127.5, 127.5), batch_size = 32, verbose = FALSE)
}
\arguments{
\item{image}{input image}

\item{model}{model object or filename see \code{getPretrainedNetwork}}

\item{strideLength}{stride length should be less than patch size}

\item{targetRange}{a vector defining min max of each patch, eg -127.5, 127.5
Output images will be scaled back to original intensity.  This range should
match the mapping used in the training of the network.}

\item{batch_size}{for prediction call}

\item{verbose}{If \code{TRUE}, show status messages}
}
\value{
image upscaled to resolution provided by network
}
\description{
Apply pretrained super-resolution network
}
\details{
This function does patch-wise pre-processing and image reconstruction given
the input super-resolution model.
}
\examples{
\dontrun{
simg <- applySuperResolutionModel( ri( 1 ), getPretrainedNetwork( "dbpn4x" ) )
}
}
\author{
Avants BB
}
