% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histology.R
\name{arterialLesionSegmentation}
\alias{arterialLesionSegmentation}
\title{Arterial lesion segmentation}
\usage{
arterialLesionSegmentation(image, verbose = FALSE)
}
\arguments{
\item{image}{input histology image.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Perform arterial lesion segmentation using U-net.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "image.nii.gz" )
output <- arterialLesionSegmentation( image )
}
}
\author{
Tustison NJ
}
