% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attentionUtilities.R
\docType{class}
\name{AttentionAugmentationLayer2D}
\alias{AttentionAugmentationLayer2D}
\title{Attention augmentation layer (2-D)}
\value{
output of tensor shape \link{batchSize, height, width, depthOfQueries}.
}
\description{
Attention augmentation layer (2-D)
}
\section{Arguments}{

\describe{
\item{depthOfQueries}{number of filters for queries.}
\item{depthOfValues}{number of filters for values.}
\item{numberOfHeads}{number of attention heads to use. It is required
that depthOfQueries/numberOfHeads > 0.}
\item{isRelative}{whether or not to use relative encodings.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
