% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attentionUtilities.R
\docType{class}
\name{AttentionLayer3D}
\alias{AttentionLayer3D}
\title{Attention layer (3-D)}
\value{
output of tensor shape.
}
\description{
Attention layer (3-D)
}
\section{Arguments}{

\describe{
\item{numberOfChannels}{number of channels.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
