% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomImageTransformAugmentation.R
\name{basisWarp}
\alias{basisWarp}
\title{Generate a deformable map from basis}
\usage{
basisWarp(deformationBasis, betaParameters, numberOfCompositions = 2,
  spatialSmoothing = 0)
}
\arguments{
\item{deformationBasis}{list containing deformationBasis set}

\item{betaParameters}{vector containing deformationBasis set parameters}

\item{numberOfCompositions}{integer greater than or equal to one}

\item{spatialSmoothing}{spatial smoothing for generated deformation}
}
\value{
list of fields
}
\description{
The function will generate a deformable transformation (via exponential map)
from a basis of deformations.
}
\examples{
\dontrun{
library( ANTsR )
i1 = ri( 1 ) \%>\% resampleImage( 4 )
i2 = ri( 2 ) \%>\% resampleImage( 4 )
reg = antsRegistration( i1, i2, 'SyN' )
w =  composeTransformsToField( i1, reg$fwd )
bw = basisWarp( list( w, w ), c( 0.25, 0.25 ), 2, 0 )
bwApp = applyAntsrTransformToImage( bw, i2, i1 )
bw = basisWarp( list( w, w ), c( 0.25, 0.25 )*(-1.0), 2, 0 ) # inverse
bwApp = applyAntsrTransformToImage( bw, i1, i2 )
}
}
\seealso{
\code{\link{randomImageTransformParametersBatchGenerator}}
}
\author{
Avants BB
}
