% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{binary_dice_coefficient}
\alias{binary_dice_coefficient}
\title{Dice function for binary segmentation problems}
\usage{
binary_dice_coefficient(y_true, y_pred, smoothingFactor = 0)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}

\item{smoothingFactor}{parameter for smoothing the metric.}
}
\value{
Dice value (negative)
}
\description{
Note:  Assumption is that y_true is \emph{not} a one-hot representation
of the segmentation batch.  For use with e.g., sigmoid activation.
}
\examples{

library( ANTsR )
library( ANTsRNet )
library( keras )

model <- createUnetModel2D( c( 64, 64, 1 ) )

dice_loss <- binary_dice_coefficient( smoothingFactor = 0.1 )

model \%>\% compile( loss = dice_loss,
  optimizer = optimizer_adam( lr = 0.0001 ) )

rm(model); gc()
}
\author{
Tustison NJ
}
