% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainMraVesselSegmentation.R
\name{brainMraVesselSegmentation}
\alias{brainMraVesselSegmentation}
\title{MRA-TOF vessel segmentation.}
\usage{
brainMraVesselSegmentation(
  mra,
  mask = NULL,
  predictionBatchSize = 16,
  patchStrideLength = 32,
  verbose = FALSE
)
}
\arguments{
\item{mra}{input mra image.}

\item{mask}{input binary mask which defines the patch extraction.
If not supplied, one is estimated.}

\item{predictionBatchSize}{Control memory usage for prediction.  More consequential
for GPU-usage.}

\item{patchStrideLength}{3-D vector or int.   Dictates the stride length for
accumulating predicting patches.}

\item{verbose}{print progress.}
}
\value{
Probability image.
}
\description{
Perform MRA-TOF vessel segmentation.  Training data taken from the
(https://data.kitware.com/#item/58a372e48d777f0721a64dc9).
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

}
}
\author{
Tustison NJ
}
