% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainTumorSegmentation.R
\name{brainTumorSegmentation}
\alias{brainTumorSegmentation}
\title{Brain tumor probabilistic segmentation}
\usage{
brainTumorSegmentation(
  flair,
  t1,
  t1Contrast,
  t2,
  predictionBatchSize = 16,
  patchStrideLength = 32,
  doPreprocessing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image (not skull-stripped).}

\item{t1}{input 3-D T1-weighted brain image (not skull-stripped).}

\item{t1Contrast}{input 3-D T1-weighted contrast brain image (not skull-stripped).}

\item{t2}{input 3-D T2-weighted brain image (not skull-stripped).}

\item{predictionBatchSize}{Control memory usage for prediction.  More consequential
for GPU-usage.}

\item{patchStrideLength}{3-D vector or int.   Dictates the stride length for
accumulating predicting patches.}

\item{doPreprocessing}{perform n4 bias correction, intensity truncation, brain
extraction.}

\item{verbose}{print progress.}
}
\value{
Brain tumor segmentation probability images (4 tumor tissue types).
}
\description{
Perform brain tumor probabilistic segmentation given pre-aligned
FLAIR, T1, T1 contrast, and T2 images.  Note that the underlying
model is 3-D and requires images to be of > 64 voxels in each
dimension.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

t1 <- antsImageRead( "t1.nii.gz" )
flair <- antsImageRead( "flair.nii.gz" )
}
}
\author{
Tustison NJ
}
