% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{categorical_focal_loss}
\alias{categorical_focal_loss}
\title{Function for categorical focal gain}
\usage{
categorical_focal_loss(y_true, y_pred, gamma = 2, alpha = 0.25)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}

\item{gamma}{focusing parameter for modulating factor (1-p). Default = 2.0.}

\item{alpha}{weighing factor in balanced cross entropy. Default = 0.25.}
}
\value{
function value
}
\description{
The negative of the categorical focal loss discussed
in this paper:
}
\details{
https://arxiv.org/pdf/1708.02002.pdf

and ported from this implementation:

https://github.com/umbertogriffo/focal-loss-keras/blob/master/losses.py

Used to handle imbalanced classes .
}
\examples{

library( ANTsRNet )
library( keras )

model <- createUnetModel2D( c( 64, 64, 1 ) )

metric_categorical_focal_gain <-
  custom_metric( "categorical_focal_gain",
    categorical_focal_gain( alpha = 0.25, gamma = 2.0 ) )

model \%>\% compile( loss = categorical_focal_loss( alpha = 0.25, gamma = 2.0 ),
 optimizer = optimizer_adam( lr = 0.0001 ),
   metrics = c( metric_categorical_focal_gain ) )

}
\author{
Tustison NJ
}
