% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerebellumMorphology.R
\name{cerebellumMorphology}
\alias{cerebellumMorphology}
\title{Cerebellum tissue segmentation, Schmahmann parcellation, and thickness.}
\usage{
cerebellumMorphology(
  t1,
  cerebellumMask = NULL,
  computeThicknessImage = FALSE,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted whole head image.}

\item{cerebellumMask}{Option for initialization.  If not specified, the
cerebellum ROI is determined using ANTsXNet brain_extraction followed by
registration to a template.}

\item{computeThicknessImage}{Compute KellyKapowski thickness image of the gray
matter.}

\item{doPreprocessing}{Perform N4 bias correction and spatiall normalize to template space.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be ressed, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
List consisting of the multiple segmentation images and probability
images for each label and foreground.  Optional thickness image.
}
\description{
Perform cerebellum segmentation using a modification of the set of
MaGET cerebellum atlases --- \url{https://www.cobralab.ca/cerebellum-lobules}.
}
\details{
The tissue labeling is as follows:
\itemize{
\item{Label 1 :}{CSF}
\item{Label 2 :}{Gray matter}
\item{Label 3 :}{White matter}
}

The parcellation labeling is as follows:
\itemize{
\item{Label 1  :}{L_I_II}
\item{Label 2  :}{L_III}
\item{Label 3  :}{L_IV}
\item{Label 4  :}{L_V}
\item{Label 5  :}{L_VI}
\item{Label 6  :}{L_Crus_I}
\item{Label 7  :}{L_Crus_II}
\item{Label 8  :}{L_VIIB}
\item{Label 9  :}{L_VIIIA}
\item{Label 10 :}{L_VIIIB}
\item{Label 11 :}{L_IX}
\item{Label 12 :}{L_X}
\item{Label 101:}{R_I_II}
\item{Label 102:}{R_III}
\item{Label 103:}{R_IV}
\item{Label 104:}{R_V}
\item{Label 105:}{R_VI}
\item{Label 106:}{R_Crus_I}
\item{Label 107:}{R_Crus_II}
\item{Label 108:}{R_VIIB}
\item{Label 109:}{R_VIIIA}
\item{Label 110:}{R_VIIIB}
\item{Label 111:}{R_IX}
\item{Label 112:}{R_X}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
cereb <- cerebellumMorphology( image )
}
}
\author{
Tustison NJ, Tustison MG
}
