% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEmbeddedClusteringUtilities.R
\docType{class}
\name{ClusteringLayer}
\alias{ClusteringLayer}
\title{Clustering layer for Deep Embedded Clustering}
\value{
clustering layer
}
\description{
Clustering layer for Deep Embedded Clustering
}
\section{Usage}{

\preformatted{outputs <- layer_clustering( numberOfClusters )}
}

\section{Arguments}{

\describe{
\item{numberOfClusters}{number of clusters.}
\item{initialClusterWeights}{}
\item{alpha}{parameter}
\item{alpha}{name}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
