% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInpaintingDeepFillModel.R
\docType{class}
\name{ContextualAttentionLayer3D}
\alias{ContextualAttentionLayer3D}
\title{Contextual attention layer (3-D)}
\value{
output tensor with the same shape as the input.
}
\description{
Contextual attention layer for generative image inpainting described in
}
\details{
Jiahui Yu, et al., Generative Image Inpainting with Contextual Attention,
CVPR 2018.

available here:\preformatted{    \code{https://arxiv.org/abs/1801.07892}
}
}
\section{Usage}{

\preformatted{layer <- ContextualAttentionLayer3D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )
}
}

\section{Arguments}{

\describe{
\item{layer}{A \code{process} object.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$build}

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\examples{
x = ContextualAttentionLayer3D$new()
x$build()
}
\author{
Tustison NJ
}
