% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConvolutionLayers.R
\name{convOutputLength}
\alias{convOutputLength}
\title{Compute output length in single dimension for convolution filter}
\usage{
convOutputLength(
  inputLength,
  filterSize,
  stride,
  dilation = 1,
  padding = c("same", "valid", "full", "causal")
)
}
\arguments{
\item{inputLength}{input size along a single dimension.}

\item{filterSize}{kernel size along a single dimension.}

\item{stride}{stride length along a single dimension.}

\item{dilation}{dilation rate along a single dimension.}

\item{padding}{type of padding.  Can be "same", "valid", "full", or "causal".}
}
\value{
the output size along a single dimension.
}
\description{
Utility function to calculate the output shape of a convolutional
filter along a single dimension.  It's based on the python keras
utility located here
}
\details{
https://github.com/keras-team/keras/blob/master/keras/utils/conv_utils.py#L118-L141

but I can't see to locate it's R analog so it's reproduced explicitly
in ANTsRNet.
}
\examples{
library( ANTsRNet )
outputSize <- convOutputLength( 256, filterSize = 3, stride = 2, padding = "same" )
testthat::expect_identical( outputSize, 128 )
}
\author{
Tustison NJ
}
