% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corticalThickness.R
\name{corticalThickness}
\alias{corticalThickness}
\title{Cortical thickness using deep learning}
\usage{
corticalThickness(t1, antsxnetCacheDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{t1}{input 3-D unprocessed T1-weighted brain image}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
Cortical thickness image and segmentation probability images.
}
\description{
Perform KellyKapowski cortical thickness using \code{deepAtropos} for
segmentation.  Description concerning implementaiton and evaluation:
}
\details{
\url{https://www.medrxiv.org/content/10.1101/2020.10.19.20215392v1}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
kk <- corticalThickness( image )
}
}
\author{
Tustison NJ
}
