% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAutoencoderModel.R
\name{createAutoencoderModel}
\alias{createAutoencoderModel}
\title{Function for creating a symmetric autoencoder model.}
\usage{
createAutoencoderModel(
  numberOfUnitsPerLayer,
  activation = "relu",
  initializer = "glorot_uniform"
)
}
\arguments{
\item{activation}{activation type for the dense layers}

\item{initializer}{initializer type for the dense layers}

\item{numberOfUnitsLayer}{vector defining the number of units
in the encoding branch}
}
\value{
two models:  the encoder and auto-encoder
}
\description{
Builds an autoencoder based on the specified array definining the
number of units in the encoding branch.  Ported to Keras R from the
Keras python implementation here:
}
\details{
\url{https://github.com/XifengGuo/DEC-keras}
}
\examples{

library( ANTsRNet )
library( keras )

ae <- createAutoencoderModel( c( 784, 500, 500, 2000, 10 ) )

}
\author{
Tustison NJ
}
