% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDeepDenoiseSuperResolutionModel.R
\name{createDeepDenoiseSuperResolutionModel3D}
\alias{createDeepDenoiseSuperResolutionModel3D}
\title{3-D implementation of the deep denoise image super resolution architecture.}
\usage{
createDeepDenoiseSuperResolutionModel3D(
  inputImageSize,
  layers = 2,
  lowestResolution = 64,
  convolutionKernelSize = c(3, 3, 3),
  poolSize = c(2, 2, 2),
  strides = c(2, 2, 2)
)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{layers}{number of architecture layers.}

\item{lowestResolution}{number of filters at the beginning and end of
the architecture.}

\item{convolutionKernelSize}{3-D vector defining the kernel size
during the encoding path}

\item{poolSize}{3-D vector defining the region for each pooling layer.}

\item{strides}{3-D vector describing the stride length in each direction.}
}
\value{
a keras model for image super resolution
}
\description{
Creates a keras model of the expanded image super resolution deep learning
framework based on the following python implementation:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    \url{https://github.com/titu1994/Image-Super-Resolution}
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
createDeepDenoiseSuperResolutionModel3D(c(256L, 256L, 45L, 1L))
}
}
\author{
Tustison NJ
}
