% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/createDenoisingAutoEncoderSuperResolutionModel.R
\name{createDenoisingAutoEncoderSuperResolutionModel3D}
\alias{createDenoisingAutoEncoderSuperResolutionModel3D}
\title{3-D implementation of the expanded image super resolution architecture.}
\usage{
createDenoisingAutoEncoderSuperResolutionModel3D(inputImageSize,
  convolutionKernelSizes = list(c(3, 3, 3), c(5, 5, 5)),
  numberOfEncodingLayers = 2, numberOfFilters = 64)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{convolutionKernelSizes}{a 2-element list of 3-D vectors specifying the
kernel size at each convolution layer.  The first element is the kernel size
of the encoding layers and the 2nd element is the kernel size of the final
convolution layer.}

\item{numberOfEncodingLayers}{the number of encoding layers.}

\item{numberOfFilters}{the number of filters for each encoding layer.}
}
\value{
a keras model for image super resolution
}
\description{
Creates a keras model of the expanded image super resolution deep learning
framework based on the following python implementation:
}
\details{
\url{https://github.com/titu1994/Image-Super-Resolution}
}
\examples{
\dontrun{ 
}
}
\author{
Tustison NJ
}
