% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDenseModel.R
\name{createDenseModel}
\alias{createDenseModel}
\title{Simple multilayer dense network.}
\usage{
createDenseModel(
  inputVectorSize,
  numberOfFiltersAtBaseLayer = 512,
  numberOfLayers = 2,
  mode = "classification",
  numberOfClassificationLabels = 1000
)
}
\arguments{
\item{inputVectorSize}{Specifies the length of the input vector.}

\item{numberOfFiltersAtBaseLayer}{Number of filters at the initial dense layer.
This number is halved for each subsequent layer.}

\item{numberOfLayers}{Number of dense layers defining the model.}

\item{mode}{'classification' or 'regression'.}

\item{numberOfClassificationLabels}{Specifies output for "classification" networks.}
}
\value{
a keras model
}
\description{
Simple multilayer dense network.
}
\examples{

library( ANTsRNet )

model <- createDenseModel(166)

}
\author{
Tustison NJ
}
