% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createHyperMapp3rUnetModel3D}
\alias{createHyperMapp3rUnetModel3D}
\title{Implementation of the "HyperMapp3r" U-net architecture}
\usage{
createHyperMapp3rUnetModel3D(inputImageSize, dataFormat = "channels_first")
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
described here:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\url{https://pubmed.ncbi.nlm.nih.gov/35088930/}
}\if{html}{\out{</div>}}

with the implementation available here:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\url{https://github.com/mgoubran/HyperMapp3r}
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{

model <- createHyperMapp3rUnetModel3D( c( 2, 224, 224, 224 ) )

}
}
\author{
Tustison NJ
}
