% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createHypothalamusUnetModel3D}
\alias{createHypothalamusUnetModel3D}
\title{3-D u-net architecture for hypothalamus segmentation}
\usage{
createHypothalamusUnetModel3D(inputImageSize)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions only
(number of channels is 1).}
}
\value{
a u-net keras model
}
\description{
Implementation of the U-net architecture for hypothalamus segmentation
described in
}
\details{
https://pubmed.ncbi.nlm.nih.gov/32853816/

and ported from the original implementation:

\if{html}{\out{<div class="sourceCode">}}\preformatted{   https://github.com/BBillot/hypothalamus_seg
}\if{html}{\out{</div>}}

The network has is characterized by the following parameters:
\itemize{
\item{3 resolution levels:  24 ---> 48 ---> 96 filters}
\item{convolution: kernel size:  c(3, 3, 3), activation: 'elu'}
\item{pool size: c(2, 2, 2)}
}
}
\examples{
# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )

model <- createHypothalamusUnetModel3d( c( 160, 160, 160 ) )

print( model )

}
\author{
Tustison NJ
}
