% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createImageSuperResolutionModel.R
\name{createImageSuperResolutionModel2D}
\alias{createImageSuperResolutionModel2D}
\title{2-D implementation of the image super resolution deep learning architecture.}
\usage{
createImageSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSizes = list(c(9, 9), c(1, 1), c(5, 5)),
  numberOfFilters = c(64, 32)
)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{convolutionKernelSizes}{a list of 2-D vectors specifying the kernel
size at each convolution layer.  Default values are the same as given in
the original paper.  The length of kernel size vectors must be 1 greater
than the vector length of the number of filters.}

\item{numberOfFilters}{a vector containing the number of filters for each
convolutional layer.  Default values are the same as given in the original
paper.}
}
\value{
a keras model for image super resolution
}
\description{
Creates a keras model of the image super resolution deep learning framework.
based on the paper available here:
}
\details{
\preformatted{    \url{https://arxiv.org/pdf/1501.00092}
}

This particular implementation is based on the following python
implementation:\preformatted{    \url{https://github.com/titu1994/Image-Super-Resolution}
}
}
\examples{
createImageSuperResolutionModel2D(c( 100, 100, 1 ))
gc()
}
\author{
Tustison NJ
}
