% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createLiWmhUnetModel2D}
\alias{createLiWmhUnetModel2D}
\title{Implementation of the Hongwei Li's U-net architecture for WMH segmentation}
\usage{
createLiWmhUnetModel2D(inputImageSize)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
described here:
}
\details{
\preformatted{\url{https://pubmed.ncbi.nlm.nih.gov/30125711/}
}

with the original implementation available here:\preformatted{\url{https://github.com/hongweilibran/wmh_ibbmTum}
}
}
\examples{
\dontrun{

model <- createLiWmhUnetModel2D( c( 200, 200, 1 ) )

}
}
\author{
Tustison NJ
}
