% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVariationalAutoencoderModel.R
\name{createMMDVariationalAutoencoderModel2D}
\alias{createMMDVariationalAutoencoderModel2D}
\title{Function for creating 2-D MMD variational autoencoder model.}
\usage{
createMMDVariationalAutoencoderModel2D(
  inputImageSize,
  latentDimension = 2L,
  numberOfEncodingLayers = 2L,
  numberOfFiltersAtBaseLayer = 64L,
  kernelSize = 4L,
  strides = 2L,
  numberOfDenseUnits = 1024L
)
}
\arguments{
\item{numberOfUnitsPerLayer}{vector defining the number of units
in the encoding branch}

\item{activation}{activation type for the dense layers}

\item{initializer}{initializer type for the dense layers}
}
\value{
two models:  the encoder and auto-encoder
}
\description{
Builds an autoencoder based on the specified array definining the
number of units in the encoding branch.  Ported to Keras R from the
Keras python implementation here:
}
\details{
\url{https://github.com/XifengGuo/DEC-keras}
}
\examples{

library( ANTsRNet )
library( keras )

ae <- createAutoencoderModel( c( 784, 500, 500, 2000, 10 ) )

}
\author{
Tustison NJ
}
