% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUnetModel.R
\name{createNoBrainerUnetModel3D}
\alias{createNoBrainerUnetModel3D}
\title{3-D implementation of the "NoBrainer" U-net architecture}
\usage{
createNoBrainerUnetModel3D(inputImageSize, numberOfOutputs = 1,
  numberOfFiltersAtBaseLayer = 16, convolutionKernelSize = c(3, 3, 3),
  deconvolutionKernelSize = c(2, 2, 2))
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).}

\item{numberOfOutputs}{Number of segmentation labels.}

\item{numberOfFiltersAtBaseLayer}{number of filters at the beginning and end
of the \verb{'U'}.  Doubles at each descending/ascending layer. Default = 16.}

\item{convolutionKernelSize}{3-d vector defining the kernel size
during the encoding path.}

\item{deconvolutionKernelSize}{3-d vector defining the kernel size
during the decoding.}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
avaialable here:
}
\details{
\url{https://github.com/neuronets/nobrainer/blob/master/nobrainer/models/unet.py}
}
\examples{

# Simple examples, must run successfully and quickly. These will be tested.

library( ANTsRNet )
library( keras )

model <- createNoBrainerUnetModel3D( c( 128, 128, 128, 1 ) )
url <- "https://github.com/neuronets/nobrainer-models/releases/download/0.1/brain-extraction-unet-128iso-weights.h5"
weightsFile <- "nobrainerWeights.h5"
download.file( url, weightsFile )
model$load_weights( weightsFile )

print( model )

}
\author{
Tustison NJ
}
