% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResNetSuperResolutionModel.R
\name{createResNetSuperResolutionModel2D}
\alias{createResNetSuperResolutionModel2D}
\title{2-D implementation of the ResNet image super resolution architecture.}
\usage{
createResNetSuperResolutionModel2D(inputImageSize,
  convolutionKernelSize = c(3, 3), numberOfFilters = 64,
  numberOfResidualBlocks = 5, numberOfResNetBlocks = 1)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{convolutionKernelSize}{a vector specifying the kernel size for
convolution.}

\item{numberOfFilters}{the number of filters for each encoding layer.}

\item{numberOfResidualBlocks}{the number of residual blocks.}

\item{numberOfResNetBlocks}{the number of resNet blocks.  Each block
will double the upsampling amount.}
}
\value{
a keras model for ResNet image super resolution
}
\description{
Creates a keras model of the expanded image super resolution deep learning
framework based on the following python implementation:
}
\details{
\url{https://github.com/titu1994/Image-Super-Resolution}
}
\examples{
\dontrun{
}
}
\author{
Tustison NJ
}
