% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomModel.R
\name{createSimpleFullyConvolutionalNeuralNetworkModel3D}
\alias{createSimpleFullyConvolutionalNeuralNetworkModel3D}
\title{Implementation of the "SCFN" architecture for Brain/Gender prediction}
\usage{
createSimpleFullyConvolutionalNeuralNetworkModel3D(
  inputImageSize,
  numberOfFiltersPerLayer = c(32, 64, 128, 256, 256, 64),
  numberOfBins = 40,
  dropoutRate = 0.5,
  doExperimentalVariant = FALSE
)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).}

\item{numberOfFiltersPerLayer}{number of filters for the convolutional layers}

\item{numberOfBins}{number of bins for final softmax output.}

\item{dropoutRate}{dropout rate before final convolution layer.}
}
\value{
a SCFN keras model
}
\description{
Creates a keras model implementation of the Simple Fully Convolutional
Network model from the FMRIB group:
}
\details{
\preformatted{    \url{https://github.com/ha-ha-ha-han/UKBiobank_deep_pretrain}
}
}
\examples{

library( ANTsRNet )

model <- createSimpleFullyConvolutionalNeuralNetworkModel3D( list( NULL, NULL, NULL, 1 ) )

}
\author{
Tustison NJ
}
