% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createSysuMediaUnetModel3D}
\alias{createSysuMediaUnetModel3D}
\title{3-D variant of the sysu_media U-net architecture}
\usage{
createSysuMediaUnetModel3D(
  inputImageSize,
  numberOfFilters = NULL,
  anatomy = "wmh"
)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.
This will be \code{c(200, 200, 2)} for t1/flair input and
\code{c(200, 200, 1)} for flair-only input for wmhs.  For the
claustrum, it is \code{c(180, 180, 1)}.}

\item{numberOfFilters}{specifies schedule for the encoding/decoding
layers.}

\item{anatomy}{"wmh" or "claustrum". Vestigial option from the original
2-D version.}
}
\value{
a u-net keras model
}
\description{
3-D variant of the sysu_media U-net architecture
}
\examples{
\dontrun{

model <- createSysuMediaUnetModel3D( c( 64, 64, 64, 2 ), numberOfFilters = c( 64, 128, 256, 512 ) )

}
}
\author{
Tustison NJ
}
