% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unetUtilities.R
\name{decodeUnet}
\alias{decodeUnet}
\title{Decoding function for the u-net prediction outcome}
\usage{
decodeUnet(yPredicted, domainImage)
}
\arguments{
\item{yPredicted}{an array of shape (\code{batchSize}, \code{width},
\code{height}, \code{<depth>}, \code{numberOfSegmentationLabels})}

\item{domainImage}{image definining the geometry of the returned probability
images.}
}
\value{
a list of list of probability images.
}
\description{
Function for translating the U-net predictions to ANTsR probability
images.
}
\author{
Tustison NJ
}
