% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unetUtilities.R
\name{decodeY}
\alias{decodeY}
\title{Decoding function for Y_predicted}
\usage{
decodeY(yPredicted, domainImage)
}
\arguments{
\item{yPredicted}{an array of shape

        `(batchSize, width, height, <depth>, numberOfSegmentationLabels)`}
}
\value{
a list of list of probability images.
}
\description{
Function for translating the U-net predictions to ANTsR probability 
images.
}
\examples{

\dontrun{ 

library( keras )

}
}
\author{
Tustison NJ
}
