% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepAtropos.R
\name{deepAtropos}
\alias{deepAtropos}
\title{Six tissue segmentation}
\usage{
deepAtropos(t1, doPreprocessing = TRUE, useSpatialPriors = 1, verbose = FALSE)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted brain image.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{useSpatialPriors}{Use MNI spatial tissue priors (0 or 1).  Currently,
only '0' (no priors) and '1' (cerebellar prior only) are the only two options.
Default is 1.}

\item{verbose}{print progress.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform Atropos-style six tissue segmentation using deep learning
}
\details{
The labeling is as follows:
\itemize{
\item{Label 0:}{background}
\item{Label 1:}{CSF}
\item{Label 2:}{gray matter}
\item{Label 3:}{white matter}
\item{Label 4:}{deep gray matter}
\item{Label 5:}{brain stem}
\item{Label 6:}{cerebellum}
}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item denoising,
\item brain extraction, and
\item affine registration to MNI.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- deepAtropos( image )
}
}
\author{
Tustison NJ
}
