% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepFlash.R
\name{deepFlash}
\alias{deepFlash}
\title{Hippocampal/Enthorhinal segmentation using "Deep Flash"}
\usage{
deepFlash(
  t1,
  t2 = NULL,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  useRankIntensity = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted brain image.}

\item{t2}{optional raw or preprocessed 3-D T2-weighted brain image.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{useRankIntensity}{If false, use histogram matching with cropped template
ROI.  Otherwise, use a rank intensity transform on the cropped ROI.}

\item{verbose}{print progress.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform hippocampal/entorhinal segmentation in T1 and T1/T2 images using
labels from Mike Yassa's lab---\url{https://faculty.sites.uci.edu/myassa/}
}
\details{
Current citation:
\url{https://www.medrxiv.org/content/10.1101/2023.01.17.23284693v4}

The labeling is as follows:
\itemize{
\item{Label 0 :}{background}
\item{Label 5 :}{left aLEC}
\item{Label 6 :}{right aLEC}
\item{Label 7 :}{left pMEC}
\item{Label 8 :}{right pMEC}
\item{Label 9 :}{left perirhinal}
\item{Label 10:}{right perirhinal}
\item{Label 11:}{left parahippocampal}
\item{Label 12:}{right parahippocampal}
\item{Label 13:}{left DG/CA2/CA3/CA4}
\item{Label 14:}{right DG/CA2/CA3/CA4}
\item{Label 15:}{left CA1}
\item{Label 16:}{right CA1}
\item{Label 17:}{left subiculum}
\item{Label 18:}{right subiculum}
}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item affine registration to deep flash template.
which is performed on the input images if \code{doPreprocessing = TRUE}.
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- deepFlash( image )
}
}
\author{
Tustison NJ
}
