% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\docType{data}
\name{.antsxnetCacheDirectory}
\alias{.antsxnetCacheDirectory}
\title{Cache directory for ANTsXNet data}
\format{
An object of class \code{character} of length 1.
}
\usage{
.antsxnetCacheDirectory
}
\description{
Cache directory for ANTsXNet data
}
\keyword{datasets}
