% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\name{drawRectangles}
\alias{drawRectangles}
\title{Plotting function for 2-D object detection visualization.}
\usage{
drawRectangles(
  image,
  boxes,
  boxColors = "red",
  confidenceValues = NULL,
  captions = NULL
)
}
\arguments{
\item{image}{standard image using something like \pkg{jpeg::readJPEG}.}

\item{boxes}{a data frame or comprising where each row has the
format: xmin, xmax, ymin, ymax.}

\item{boxColors}{Optional scalar or vector of length = \code{numberOfBoxes}
used for determining the colors of the different boxes.}

\item{confidenceValues}{Optional vector of length = \code{numberOfBoxes} where
each element is in the range \verb{[0, 1]}.  Used for determining border width.}

\item{captions}{Optional vector of length = \code{numberOfBoxes} where
each element is the caption rendered with each box.}
}
\description{
Renders boxes on objects within rasterized images.
}
\author{
Tustison NJ
}
