% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histology.R
\name{e13x5BrainExtraction}
\alias{e13x5BrainExtraction}
\title{Perform brain extraction of Allen's E13.5 mouse embroyonic data.}
\usage{
e13x5BrainExtraction(
  image,
  view = c("coronal", "sagittal"),
  whichAxis = 3,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input image}

\item{view}{Two trained networks are available:  "coronal" or "sagittal".}

\item{whichAxis}{If 3-D image, which_axis specifies the direction of the "view".}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Perform brain extraction of Allen's E13.5 mouse embroyonic data.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "image.nii.gz" )
output <- e13x5BrainExtraction( image )
}
}
\author{
Tustison NJ
}
