% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungSegmentation.R
\name{elBicho}
\alias{elBicho}
\title{Functional lung segmentation.}
\usage{
elBicho(
  ventilationImage,
  mask,
  useCoarseSlicesOnly = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ventilationImage}{input ventilation image}

\item{mask}{input mask image}

\item{useCoarseSlicesOnly}{if \code{TRUE}, apply network only in the
dimension of greatest slice thickness.  If \code{FALSE}, apply to all
dimensions and average the results.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
ventilation segmentation and corresponding probability images
}
\description{
Perform functional lung segmentation using hyperpolarized gases.
}
\details{
\url{https://pubmed.ncbi.nlm.nih.gov/30195415/}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "flair.nii.gz" )
probabilityMask <-sysuMediaWmhSegmentation( image )
}
}
\author{
Tustison NJ
}
