% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unetUtilities.R
\name{encodeUnet}
\alias{encodeUnet}
\title{One-hot encoding function}
\usage{
encodeUnet(segmentationsArray, segmentationLabels = NULL)
}
\arguments{
\item{segmentationsArray}{an array of shape (\code{batchSize}, \code{width},
\code{height}, \code{<depth>})}

\item{segmentationLabels}{vector of segmentation labels.  Note that a
background label (typically 0) needs to be included.}
}
\value{
an n-D array of shape
\eqn{ batchSize \times width \times height \times <depth> \times numberOfSegmentationLabels }
}
\description{
Function for translating the segmentations to a one-hot representation.
}
\author{
Tustison NJ
}
