% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unetUtilities.R
\name{encodeY}
\alias{encodeY}
\title{Encoding function for Y_train}
\usage{
encodeY(groundTruthSegmentations, segmentationLabels)
}
\arguments{
\item{groundTruthSegmentations}{an array of shape\preformatted{  (batchSize, width, height, <depth>)
}}

\item{segmentationLabels}{vector of segmentation labels.  Note that a
background label (typically '0') needs to be included.}
}
\value{
an n-D array of shape\preformatted{    `(batchSize, width, height, <depth>, numberOfSegmentationLabels)`
}
}
\description{
Function for translating the segmentations to something readable by the
optimization process.
}
\examples{

\dontrun{ 

library( keras )

}
}
\author{
Tustison NJ
}
