% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{error}
\alias{error}
\alias{MSE}
\alias{MAE}
\alias{PSNR}
\title{Error measurements of images (from redr)}
\usage{
MSE(x, y = NULL)

MAE(x, y = NULL)

PSNR(x, y)
}
\arguments{
\item{x, y}{\code{antsImage} objects}
}
\description{
This function calculates error between two images
}
\section{Functions}{
\itemize{
\item \code{MSE}: Mean Squared Error

\item \code{MAE}: Mean Absolute Error

\item \code{PSNR}: Peak Signal-to-Noise Ratio
}}

\examples{
library( ANTsRCore )
MSE( ri(1), ri(2) )
# alternatively it can be done like:
MSE( ri(1) - ri(2) )
}
