% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianDiffusionUtilities.R
\docType{class}
\name{GaussianDiffusion}
\alias{GaussianDiffusion}
\title{Custom Gaussian Diffusion}
\description{
Adapted from
}
\details{
https://github.com/keras-team/keras-io/blob/master/examples/generative/ddpm.py
}
\section{Arguments}{

\describe{
\item{betaStart}{}
\item{betaEnd}{}
\item{timeSteps}{}
\item{clipMin}{}
\item{clipMax}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
