% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\name{getAntsxnetCacheDirectory}
\alias{getAntsxnetCacheDirectory}
\title{Get ANTsXNet Cache Directory}
\usage{
getAntsxnetCacheDirectory()
}
\value{
A character string specifying the path to the cache directory.
}
\description{
This function returns the current cache directory used by ANTsXNet to store
downloaded templates and model weights.
}
\examples{
\dontrun{
cache_dir <- getAntsxnetCacheDirectory()
print(cache_dir)
}
}
