% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\name{getANTsXNetData}
\alias{getANTsXNetData}
\title{getANTsXNetData}
\usage{
getANTsXNetData(
  fileId = c("show", "biobank", "croppedMni152", "croppedMni152Priors",
    "deepFlashPriors", "deepFlashTemplateT1", "deepFlashTemplateT1SkullStripped",
    "deepFlashTemplateT2", "deepFlashTemplateT2SkullStripped",
    "deepFlashTemplate2T1SkullStripped", "deepFlashTemplate2Labels", "mprage_hippmapp3r",
    "protonLungTemplate", "ctLungTemplate", "luna16LungPriors", "xrayLungPriors",
    "priorDktLabels", "priorDeepFlashLeftLabels", "priorDeepFlashRightLabels",
    "protonLobePriors", "S_template3", "adni", "ixi", "kirby", "mni152", 
     "nki",
    "nki10", "oasis", "magetTemplate", "magetTemplateBrainMask",
    "magetCerebellumTemplate", "magetCerebellumTemplatePriors",
    "magetCerebellumxTemplate0GenericAffine", "mraTemplate", "mraTemplateBrainMask",
    "mraTemplateVesselPrior", "bsplineT2MouseTemplate",
    "bsplineT2MouseTemplateBrainMask", "DevCCF_P56_MRI_T2_50um",
    "DevCCF_P56_MRI_T2_50um_BrainParcellationNickMask",
    "DevCCF_P56_MRI_T2_50um_BrainParcellationTctMask", "DevCCF_P04_STPT_50um",
    "DevCCF_P04_STPT_50um_BrainParcellationJayMask", 
     "hcpyaT1Template",
    "hcpyaT2Template", "hcpyaFATemplate", "hcpyaTemplateBrainMask",
    "hcpyaTemplateBrainSegmentation"),
  targetFileName
)
}
\arguments{
\item{fileId}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{targetFileName}{optional target filename}
}
\value{
filename string
}
\description{
Download data such as prefabricated templates and spatial priors.
}
\note{
See \url{https://figshare.com/authors/Nick_Tustison/441144}
or \url{https://figshare.com/authors/Brian_Avants/418551}
for some more descriptions
}
\examples{
\dontrun{
net <- getANTsXNetData("biobank")
}
}
\author{
Avants BB
}
