% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{getMixtureDensityLossFunction}
\alias{getMixtureDensityLossFunction}
\title{Returns a loss function for the mixture density.}
\usage{
getMixtureDensityLossFunction(outputDimension, numberOfMixes)
}
\arguments{
\item{outputDimension}{output dimension}

\item{numberOfMixes}{number of mixture components}
}
\value{
a function providing the mean square error accuracy
}
\description{
Ported from:
}
\details{
https://github.com/cpmpercussion/keras-mdn-layer/
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
