% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{getMixtureDensityMseAccuracyFunction}
\alias{getMixtureDensityMseAccuracyFunction}
\title{Returns a MSE accuracy function for the mixture density.}
\usage{
getMixtureDensityMseAccuracyFunction(outputDimension, numberOfMixtures)
}
\arguments{
\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of mixture components}
}
\value{
a function providing the mean square error accuracy
}
\description{
Ported from:
}
\details{
\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
