% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPretrainedNetwork.R
\name{getPretrainedNetwork}
\alias{getPretrainedNetwork}
\title{getPretrainedNetwork}
\usage{
getPretrainedNetwork(fileId, targetFileName)
}
\arguments{
\item{fileId}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{targetFileName}{optional target filename}
}
\value{
filename string
}
\description{
Downloads pretrained network/weights.
}
\examples{
\dontrun{
net <- getPretrainedNetwork( "dbpn4x" )
}
}
\author{
Avants BB
}
