% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{GMSD}
\alias{GMSD}
\title{Gradient Magnitude Similarity Deviation}
\usage{
GMSD(x, y)
}
\arguments{
\item{x}{input image.}

\item{y}{input image.}
}
\value{
scalar
}
\description{
A fast and simple metric that correlates to perceptual quality
}
\examples{

library( ANTsR )

r16 <- antsImageRead( getANTsRData( 'r16' ) )
r85 <- antsImageRead( getANTsRData( 'r85' ) )
value <- GMSD( r16, r85 )

}
\author{
Avants BB
}
