% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvardOxfordAtlasLabeling.R
\name{harvardOxfordAtlasLabeling}
\alias{harvardOxfordAtlasLabeling}
\title{Subcortical and cerebellar labeling from a T1 image.}
\usage{
harvardOxfordAtlasLabeling(t1, doPreprocessing = TRUE, verbose = FALSE)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted brain image.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{verbose}{print progress.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform HOA labeling using deep learning and data from
Morphology" number: "NIH NIMH 5R01MH112748-04". Repository:
https://github.com/HOA-2/SubcorticalParcellations'
}
\details{
The labeling is as follows:
\itemize{
\item{Label 1:}{Lateral Ventricle Left}
\item{Label 2:}{Lateral Ventricle Right}
\item{Label 3:}{CSF}
\item{Label 4:}{Third Ventricle}
\item{Label 5:}{Fourth Ventricle}
\item{Label 6:}{5th Ventricle}
\item{Label 7:}{Nucleus Accumbens Left}
\item{Label 8:}{Nucleus Accumbens Right}
\item{Label 9:}{Caudate Left}
\item{Label 10:}{Caudate Right}
\item{Label 11:}{Putamen Left}
\item{Label 12:}{Putamen Right}
\item{Label 13:}{Globus Pallidus Left}
\item{Label 14:}{Globus Pallidus Right}
\item{Label 15:}{Brainstem}
\item{Label 16:}{Thalamus Left}
\item{Label 17:}{Thalamus Right}
\item{Label 18:}{Inferior Horn of the Lateral Ventricle Left}
\item{Label 19:}{Inferior Horn of the Lateral Ventricle Right}
\item{Label 20:}{Hippocampal Formation Left}
\item{Label 21:}{Hippocampal Formation Right}
\item{Label 22:}{Amygdala Left}
\item{Label 23:}{Amygdala Right}
\item{Label 24:}{Optic Chiasm}
\item{Label 25:}{VDC Anterior Left}
\item{Label 26:}{VDC Anterior Right}
\item{Label 27:}{VDC Posterior Left}
\item{Label 28:}{VDC Posterior Right}
\item{Label 29:}{Cerebellar Cortex Left}
\item{Label 30:}{Cerebellar Cortex Right}
\item{Label 31:}{Cerebellar White Matter Left}
\item{Label 32:}{Cerebellar White Matter Right}
}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item brain extraction, and
\item affine registration to HCP.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- harvardOxfordAtlasLabeling( image )
}
}
\author{
Tustison NJ
}
