% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippMapp3rSegmentation.R
\name{hippMapp3rSegmentation}
\alias{hippMapp3rSegmentation}
\title{hippMapp3rSegmentation}
\usage{
hippMapp3rSegmentation(
  t1,
  doPreprocessing = TRUE,
  outputDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}

\item{image}{input 3-D T1-weighted brain image.}
}
\value{
labeled hippocampal mask (ANTsR image)
}
\description{
Perform HippMapp3r (hippocampal) segmentation described in
}
\details{
https://www.ncbi.nlm.nih.gov/pubmed/31609046

with models and architecture ported from

https://github.com/mgoubran/HippMapp3r

Additional documentation and attribution resources found at

https://hippmapp3r.readthedocs.io/en/latest/

Preprocessing consists of:
\itemize{
\item n4 bias correction and
\item brain extraction.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( ANTsRNet )
library( keras )

url <- "https://github.com/mgoubran/HippMapp3r/blob/master/data/test_case/mprage.nii.gz?raw=true"
imageFile <- "head.nii.gz"
download.file( url, imageFile )
image <- antsImageRead( imageFile )
imageN4 <- n4BiasFieldCorrection( image, verbose = TRUE )
segmentation <- hippMapp3rSegmentation( imageN4, verbose = TRUE )
}
}
\author{
Tustison NJ
}
