% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{hyperMapp3rSegmentation}
\alias{hyperMapp3rSegmentation}
\title{hyperMapp3rSegmentation}
\usage{
hyperMapp3rSegmentation(
  t1,
  flair,
  doPreprocessing = TRUE,
  numberOfMonteCarloIterations = 30,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{t1}{input 3-D t1-weighted MR image.  Assumed to be aligned with the flair.}

\item{flair}{input 3-D flair MR image.  Assumed to be aligned with the t1.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
white matter hyperintensity probability mask
}
\description{
Perform HyperMapp3r (white matter hyperintensity) segmentation described in
}
\details{
https://pubmed.ncbi.nlm.nih.gov/35088930/

with models and architecture ported from

https://github.com/AICONSlab/HyperMapp3r

Additional documentation and attribution resources found at

https://hypermapp3r.readthedocs.io/en/latest/

Preprocessing consists of:
\itemize{
\item n4 bias correction and
\item brain extraction.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

wmh <- hyperMapp3rSegmentation( image, verbose = TRUE )
}
}
\author{
Tustison NJ
}
