% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothalamusSegmentation.R
\name{hypothalamusSegmentation}
\alias{hypothalamusSegmentation}
\title{Hypothalamus segmentation}
\usage{
hypothalamusSegmentation(t1, antsxnetCacheDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{t1}{input 3-D T1-weighted brain image.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
hypothalamic subunit segmentation and probability images
\itemize{
\item{Label 1:  left anterior-inferior}
\item{Label 2:  left anterior-superior}
\item{Label 3:  left posterior}
\item{Label 4:  left tubular inferior}
\item{Label 5:  left tubular superior}
\item{Label 6:  right anterior-inferior}
\item{Label 7:  right anterior-superior}
\item{Label 8:  right posterior}
\item{Label 9:  right tubular inferior}
\item{Label 10: right tubular superior}
}
}
\description{
Described here:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\url{https://pubmed.ncbi.nlm.nih.gov/32853816/}
}\if{html}{\out{</div>}}

with the implementation available at:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\url{https://github.com/BBillot/hypothalamus_seg}
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
hypo <- hypothalamusSegmentation( image )
}
}
\author{
Tustison NJ
}
