% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInpaintingDeepFillModel.R,
%   R/ssdUtilities.R
\docType{class}
\name{L2NormalizationLayer2D}
\alias{L2NormalizationLayer2D}
\title{Contextual attention layer (2-D)}
\value{
output tensor with the same shape as the input.

output tensor with the same shape as the input.

output tensor with the same shape as the input.
}
\description{
Contextual attention layer for generative image inpainting described in

Contextual attention layer for generative image inpainting described in

L2 2-D normalization layer for SSD300/512 architecture described in
}
\details{
Jiahui Yu, et al., Generative Image Inpainting with Contextual Attention,
CVPR 2018.

available here:

\code{https://arxiv.org/abs/1801.07892}

Jiahui Yu, et al., Generative Image Inpainting with Contextual Attention,
CVPR 2018.

available here:

\code{https://arxiv.org/abs/1801.07892}

Wei Liu, Andrew Rabinovich, and Alexander C. Berg.  ParseNet: Looking Wider
to See Better.

available here:

\code{https://arxiv.org/abs/1506.04579}
}
\section{Usage}{

\preformatted{layer <- ContextualAttentionLayer2D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )
}


\preformatted{layer <- ContextualAttentionLayer2D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )
}


\preformatted{layer <- L2NormalizationLayer2D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )
}
}

\section{Arguments}{

\describe{
\item{layer}{A \code{process} object.}
\item{scale}{feature scale.  Default = 20}
\item{x}{}
\item{mask}{}
\item{input_shape}{}
}


\describe{
\item{layer}{A \code{process} object.}
}


\describe{
\item{layer}{A \code{process} object.}
\item{scale}{feature scale.  Default = 20}
\item{x}{}
\item{mask}{}
\item{input_shape}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$build}

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.


\code{$initialize} instantiates a new class.

\code{$build}

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.


\code{$initialize} instantiates a new class.

\code{$build}

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ

Tustison NJ

Tustison NJ
}
