% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{class}
\name{L2NormalizationLayer3D}
\alias{L2NormalizationLayer3D}
\title{L2 3-D normalization layer for SSD300/512 architecture.}
\value{
output tensor with the same shape as the input.
}
\description{
L2 3-D normalization layer for SSD300/512 architecture described in
}
\details{
Wei Liu, Andrew Rabinovich, and Alexander C. Berg.  ParseNet: Looking Wider
to See Better.

available here:

\code{https://arxiv.org/abs/1506.04579}
}
\section{Usage}{

\preformatted{layer <- L2NormalizationLayer3D$new( scale )

layer$call( x, mask = NULL )
layer$build( input_shape )
layer$compute_output_shape( input_shape )
}
}

\section{Arguments}{

\describe{
\item{layer}{A \code{process} object.}
\item{scale}{feature scale.  Default = 20}
\item{x}{}
\item{mask}{}
\item{input_shape}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$build}

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
