% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customActivationLayers.R
\name{layer_activation_log_softmax}
\alias{layer_activation_log_softmax}
\title{Log softmax layer}
\usage{
layer_activation_log_softmax(object, axis = -1, trainable = TRUE)
}
\arguments{
\item{axis}{Integer specifying which axis.}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\value{
a keras layer tensor
}
\description{
Creates a log softmax layer
}
\author{
Tustison NJ
}
