% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\name{layer_anchor_box_3d}
\alias{layer_anchor_box_3d}
\title{Anchor box layer (3-D)}
\usage{
layer_anchor_box_3d(
  object,
  imageSize,
  scale,
  nextScale,
  aspectRatios,
  variances,
  name = NULL,
  trainable = TRUE
)
}
\arguments{
\item{imageSize}{size of the image}

\item{scale}{box scale}

\item{nextScale}{box scale}

\item{aspectRatios}{list of ratios used for the boxes}

\item{variances}{list}
}
\value{
a keras layer tensor
}
\description{
Wraps a custom layer for the SSD network
}
