% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInpaintingDeepFillModel.R
\name{layer_contextual_attention_2d}
\alias{layer_contextual_attention_2d}
\title{Contextual attention layer (2-D)}
\usage{
layer_contextual_attention_2d(
  object,
  kernelSize = 3L,
  stride = 1L,
  dilationRate = 1L,
  fusionKernelSize = 0L,
  name = NULL,
  trainable = FALSE
)
}
\arguments{
\item{kernelSize}{integer specifying convolution size}

\item{stride}{integer for specifyingstride length for sampling the tensor}

\item{dilationRate}{ingeger specifying dilation}

\item{fusionKernelSize}{Enhance saliency of large patches}
}
\value{
a keras layer tensor
}
\description{
Contextual attention layer for generative image inpainting described in
}
\details{
Jiahui Yu, et al., Generative Image Inpainting with Contextual Attention,
CVPR 2018.

available here:\preformatted{    \code{https://arxiv.org/abs/1801.07892}
}
}
