% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{layer_mixture_density}
\alias{layer_mixture_density}
\title{Mixture density layer}
\usage{
layer_mixture_density(
  object,
  outputDimension,
  numberOfMixtures,
  trainable = TRUE
)
}
\arguments{
\item{object}{Object to compose layer with. This is either a
\link[keras:keras_model_sequential]{keras::keras_model_sequential} to add the layer to,
or another Layer which this layer will call.}

\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of Gaussians used to model the function}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\value{
a keras layer tensor
}
\description{
Wraps a custom mixture density layer.
}
\examples{
\dontrun{
model <- keras_model_sequential()
input_shape = c(20, 20, 1)
model = model \%>\%
  layer_conv_2d(filters = 32, kernel_size = c(3,3), activation = 'relu',
                input_shape = input_shape)
model \%>\%
  layer_mixture_density(outputDimension = c(18, 18),
  numberOfMixtures = 3)
 }
}
