% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConvolutionLayers.R
\name{layer_partial_conv_3d}
\alias{layer_partial_conv_3d}
\title{Partial convolution layer 3D}
\usage{
layer_partial_conv_3d(
  object,
  filters,
  kernelSize,
  strides = c(1L, 1L, 1L),
  padding = "valid",
  dataFormat = "channels_last",
  dilationRate = c(1L, 1L, 1L),
  activation = NULL,
  kernelInitializer = "glorot_uniform",
  biasInitializer = "zeros",
  kernelRegularizer = NULL,
  biasRegularizer = NULL,
  activityRegularizer = NULL,
  useBias = TRUE,
  trainable = TRUE
)
}
\arguments{
\item{object}{Object to compose layer with. This is either a
\link[keras:keras_model_sequential]{keras::keras_model_sequential} to add the layer to,
or another Layer which this layer will call.}

\item{filters}{number of filters}

\item{kernelSize}{kernel size}

\item{strides}{strides}

\item{padding}{padding}

\item{dataFormat}{format}

\item{dilationRate}{dilate rate}

\item{activation}{activation}

\item{kernelInitializer}{kernel initializer}

\item{biasInitializer}{bias initializer}

\item{kernelRegularizer}{kernel regularizer}

\item{biasRegularizer}{bias regularizer}

\item{activityRegularizer}{activity regularizer}

\item{useBias}{use bias}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\value{
a keras layer tensor
}
\description{
Creates an 3D partial convolution layer
}
\author{
Tustison NJ
}
