% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConvolutionLayers.R
\name{layer_partial_convolution_2d}
\alias{layer_partial_convolution_2d}
\title{Partial convolution layer 2D}
\usage{
layer_partial_convolution_2d(
  object,
  filters,
  kernel_size,
  strides = c(1L, 1L),
  padding = "valid",
  data_format = "channels_last",
  dilation_rate = c(1L, 1L),
  activation = NULL,
  kernel_initializer = "glorot_uniform",
  bias_initializer = "zeros",
  kernel_regularizer = NULL,
  bias_regularizer = NULL,
  activity_regularizer = NULL,
  use_bias = TRUE,
  trainable = TRUE
)
}
\arguments{
\item{object}{Object to compose layer with. This is either a
\link[keras:keras_model_sequential]{keras::keras_model_sequential} to add the layer to,
or another Layer which this layer will call.}

\item{filters}{number of filters}

\item{kernel_size}{kernel size}

\item{strides}{strides}

\item{padding}{padding}

\item{data_format}{format}

\item{dilation_rate}{dilate rate}

\item{activation}{activation}

\item{kernel_initializer}{kernel initializer}

\item{bias_initializer}{bias initializer}

\item{kernel_regularizer}{kernel regularizer}

\item{bias_regularizer}{bias regularizer}

\item{activity_regularizer}{activity regularizer}

\item{use_bias}{use bias}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\value{
a keras layer tensor
}
\description{
Creates an 2D partial convolution layer
}
\author{
Tustison NJ
}
