% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTensorUtilities.R
\name{layer_resample_tensor_3d}
\alias{layer_resample_tensor_3d}
\title{Resampling a spatial tensor (3-D).}
\usage{
layer_resample_tensor_3d(object, shape,
  interpolationType = "nearestNeighbor", name = NULL,
  trainable = FALSE)
}
\arguments{
\item{shape}{vector or list of length 3 specifying the shape of the output tensor.}

\item{interpolationType}{type of interpolation for resampling.  Can be
\code{nearestNeighbor}, \code{linear}, or \code{cubic}.}

\item{inputTensor}{tensor to be resampled.}
}
\value{
a keras layer tensor
}
\description{
Resamples a spatial tensor based on the specified shape and interpolation type.
}
