% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransformerNetworkUtilities.R
\name{layer_spatial_transformer_3d}
\alias{layer_spatial_transformer_3d}
\title{spatial transformer layer (3-D)}
\usage{
layer_spatial_transformer_3d(objects, resampledSize,
  transformType = "affine", interpolatorType = "linear", name = NULL)
}
\arguments{
\item{resampledSize}{size of the output in voxels}

\item{transformType}{the spatial transform}

\item{interpolatorType}{interpolation used for the sampling}
}
\value{
a keras layer tensor
}
\description{
Wraps a custom spatial transformer layer
}
