% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesionSegmentation.R
\name{lesionSegmentation}
\alias{lesionSegmentation}
\title{wholeHeadInpainting}
\usage{
lesionSegmentation(
  t1,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{t1}{input 3-D T1 brain image (not skull-stripped).}

\item{doPreprocessing}{Perform n4 bias correction, intensity truncation, brain
extraction.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
inpainted image
}
\description{
Perform in-painting for whole-head MRI
}
\author{
Tustison NJ
}
