% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDBPN4x.R
\name{linMatchIntensity}
\alias{linMatchIntensity}
\title{linMatchIntensity}
\usage{
linMatchIntensity(fromImg, toImg, polyOrder = 1, truncate = TRUE, mask)
}
\arguments{
\item{fromImg}{image whose intensity function we will match to the \code{toImg}}

\item{toImg}{defines the reference intensity function.}

\item{polyOrder}{of polynomial fit.  default is none or just linear fit.}

\item{truncate}{boolean which turns on/off the clipping of intensity.}

\item{mask}{mask the matching region}
}
\value{
the \code{fromImg} matched to the \code{toImg}
}
\description{
regression between two image intensity spaces
}
\examples{
library(ANTsRCore)
sourceImage <- antsImageRead( getANTsRData( "r16" ) )
referenceImage <- antsImageRead( getANTsRData( "r64" ) )
matchedImage <- linMatchIntensity( sourceImage, referenceImage )
}
\author{
Avants BB
}
