% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customActivationLayers.R
\docType{class}
\name{LogSoftmaxLayer}
\alias{LogSoftmaxLayer}
\title{Creates a log softmax layer}
\value{
a log softmax layer
}
\description{
Creates a log softmax layer taken from
}
\details{
\url{https://github.com/tensorflow/tensorflow/pull/25514/files}
}
\section{Arguments}{

\describe{
\item{axis}{Integer specifying the axis.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.
\code{$call} main body.
\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
