% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corticalThickness.R
\name{longitudinalCorticalThickness}
\alias{longitudinalCorticalThickness}
\title{Longitudinal cortical thickness using deep learning}
\usage{
longitudinalCorticalThickness(
  t1s,
  initialTemplate = "oasis",
  numberOfIterations = 1,
  refinementTransform = "antsRegistrationSyNQuick[a]",
  verbose = FALSE
)
}
\arguments{
\item{t1s}{input list of 3-D unprocessed T1-weighted brain images from a single subject}

\item{initialTemplate}{input image to define the orientation of the SST.  Can be a string
(see \code{getANTsXNetData}) or a specified template.  This allows the user to create a
SST outside of this routine.}

\item{numberOfIterations}{Defines the number of iterations for refining the SST.}

\item{refinementTransform}{Transform for defining the refinement registration transform.
See options in \code{antsRegistration}.}

\item{verbose}{print progress.}
}
\value{
List consisting of the SST, and a (sub-)list for each subject consisting of
the preprocessed image, cortical thickness image, segmentation probability images,
and affine mapping to the SST.
}
\description{
Perform KellyKapowski cortical thickness longitudinally using \code{deepAtropos}
for segmentation of the derived single-subject template.  It takes inspiration from
the work described here:
}
\details{
\url{https://pubmed.ncbi.nlm.nih.gov/31356207/}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
kk <- corticalThickness( image )
}
}
\author{
Tustison NJ, Avants BB
}
