% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customLosses.R
\name{loss_peak_signal_to_noise_ratio_error}
\alias{loss_peak_signal_to_noise_ratio_error}
\title{Peak-signal-to-noise ratio.}
\usage{
loss_peak_signal_to_noise_ratio_error(y_true, y_pred)
}
\arguments{
\item{y_true}{true encoded labels}

\item{y_pred}{predicted encoded labels}
}
\description{
Peak-signal-to-noise ratio.
}
