% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customLosses.R
\name{loss_pearson_correlation_coefficient_error}
\alias{loss_pearson_correlation_coefficient_error}
\title{Pearson correlation coefficient}
\usage{
loss_pearson_correlation_coefficient_error(y_true, y_pred)
}
\arguments{
\item{y_true}{true encoded labels}

\item{y_pred}{predicted encoded labels}
}
\description{
Pearson correlation coefficient
}
