% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{data}
\name{lossSsd}
\alias{lossSsd}
\title{Loss function for the SSD deep learning architecture.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
lossSsd
}
\arguments{
\item{dimension}{Image dimensionality.}

\item{backgroundRatio}{The maximum ratio of background to foreround
for weighting in the loss function.  Is rounded to the nearest integer.
Default is '3'.}

\item{minNumberOfBackgroundBoxes}{The minimum number of background boxes
to use in loss computation \emph{per batch}.  Should reflect a value in
proportion to the batch size.  Default is 0.}

\item{alpha}{Weighting factor for the localization loss in total loss
computation.}

\item{numberOfClassificationLabels}{number of classes including background.}
}
\value{
an SSD loss function
}
\description{
Creates an R6 class object for use with the SSD deep learning architecture
based on the paper
}
\details{
W. Liu, D. Anguelov, D. Erhan, C. Szegedy, S. Reed, C-Y. Fu, A. Berg.
SSD: Single Shot MultiBox Detector.

available here:

https://arxiv.org/abs/1512.02325

This particular implementation was heavily influenced by the following
python and R implementations:

https://github.com/pierluigiferrari/ssd_keras/blob/master/keras_loss_function/keras_ssd_loss.py
https://github.com/gsimchoni/ssdkeras/blob/master/R/ssd_loss.R
}
\author{
Tustison NJ
}
\keyword{datasets}
