% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungSegmentation.R
\name{lungAirwaySegmentation}
\alias{lungAirwaySegmentation}
\title{Lung airway segmentation.}
\usage{
lungAirwaySegmentation(
  ct,
  lungMask = NULL,
  predictionBatchSize = 16,
  patchStrideLength = 32,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ct}{input 3-D ct image.}

\item{lungMask}{input binary lung mask which defines the patch extraction
(label 1 = left lung, label 2 = right lung, label 3 = main airway).
If not supplied, one is estimated.}

\item{predictionBatchSize}{Control memory usage for prediction.  More consequential
for GPU-usage.}

\item{patchStrideLength}{3-D vector or int.   Dictates the stride length for
accumulating predicting patches.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
Probability image.
}
\description{
Perform pulmonary artery segmentation.  Training data taken from the
EXACT09 challenge (Lo, Pechin, et al. "Extraction of airways from CT
(EXACT'09)." https://pubmed.ncbi.nlm.nih.gov/22855226/)
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

}
}
\author{
Tustison NJ
}
