% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungExtraction.R
\name{lungExtraction}
\alias{lungExtraction}
\title{Lung extraction}
\usage{
lungExtraction(
  image,
  modality = c("proton", "protonLobes", "maskLobes", "ct", "ventilation", "xray"),
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D lung image.}

\item{modality}{image type.  Options include "proton", "ct", or "ventilation".}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Perform lung extraction.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "lung.nii.gz" )
output <- lungExtraction( image, modality = "proton" )
}
}
\author{
Tustison NJ
}
