% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{masked_mse_error}
\alias{masked_mse_error}
\title{Function for masked mean-squared error}
\usage{
masked_mse_error(y_true, y_pred, maskExcludeValue = 1)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
masked mse value
}
\description{
Function for masked mean-squared error
}
\author{
Tustison NJ
}
