% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{maximum_mean_discrepancy}
\alias{maximum_mean_discrepancy}
\title{Function for maximum-mean discrepancy}
\usage{
maximum_mean_discrepancy(y_true, y_pred, sigma = 1)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
mmd value
}
\description{
\url{https://jmlr.csail.mit.edu/papers/volume13/gretton12a/gretton12a.pdf}
}
\author{
Tustison NJ
}
