% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{metric_multilabel_dice_coefficient}
\alias{metric_multilabel_dice_coefficient}
\title{Dice metric for multilabel problems}
\usage{
metric_multilabel_dice_coefficient(...)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
Dice value
}
\description{
Dice metric for multilabel problems
}
\author{
Tustison NJ
}
