% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{metric_peak_signal_to_noise_ratio}
\alias{metric_peak_signal_to_noise_ratio}
\title{Metric to calculate peak-signal-to-noise ratio.}
\usage{
metric_peak_signal_to_noise_ratio(...)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
PSNR value
}
\description{
Metric to calculate peak-signal-to-noise ratio.
}
\author{
Tustison NJ
}
