% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{metric_pearson_correlation_coefficient}
\alias{metric_pearson_correlation_coefficient}
\title{Metric for Pearson correlation coefficient.}
\usage{
metric_pearson_correlation_coefficient(...)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
Correlation
}
\description{
Metric for Pearson correlation coefficient.
}
\author{
Tustison NJ
}
