% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\docType{class}
\name{MixtureDensityNetworkLayer}
\alias{MixtureDensityNetworkLayer}
\title{Mixture density network layer}
\value{
.
}
\description{
Mixture density network layer
}
\section{Usage}{

\preformatted{outputs <- layer_mixture_density( inputs, resampledSize )}
}

\section{Arguments}{

\describe{
\item{outputDimensino}{}
\item{numberOfMixtures}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
