% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.R
\name{mouseBrainExtraction}
\alias{mouseBrainExtraction}
\title{Mouse brain extraction}
\usage{
mouseBrainExtraction(
  image,
  modality = c("t2", "ex5coronal", "ex5sagittal"),
  returnIsotropicOutput = FALSE,
  whichAxis = 2,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D brain image (or list of images for multi-modal scenarios).}

\item{modality}{image type.  Options include:
\itemize{
\item{"t2": }{T2-weighted MRI}
\item{"ex5coronal": }{E13.5 and E15.5 mouse embroyonic histology data.}
\item{"ex5sagittal": }{E13.5 and E15.5 mouse embroyonic histology data.}
}}

\item{returnIsotropicOutput}{The network actually learns an interpolating
function specific to the mouse brain.  Setting this to true, the output
images are returned isotropically resampled.}

\item{whichAxis}{Specify direction for ex5 modalities..}

\item{verbose}{print progress.}
}
\value{
brain probability mask
}
\description{
Perform brain extraction of mouse MRI
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "brain.nii.gz" )
probabilityMask <- mouseBrainExtraction( image, modality = "t2" )
}
}
\author{
Tustison NJ
}
