% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.R
\name{mouseCorticalThickness}
\alias{mouseCorticalThickness}
\title{Mouse brain cortical thickness using deep learning}
\usage{
mouseCorticalThickness(
  t2,
  mask = NULL,
  returnIsotropicOutput = FALSE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{t2}{input 3-D unprocessed T2-weighted mouse brain image}

\item{mask}{Brain mask.  If not specified, one is estimated using ANTsXNet mouse brain
extraction.}

\item{returnIsotropicOutput}{The network actually learns an interpolating
function specific to the mouse brain.  Setting this to true, the output
images are returned isotropically resampled.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
Cortical thickness image and segmentation probability images.
}
\description{
Perform KellyKapowski cortical thickness
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
kk <- corticalThickness( image )
}
}
\author{
Tustison NJ
}
