% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mriSuperResolution.R
\name{mriSuperResolution}
\alias{mriSuperResolution}
\title{Super-resolution for MRI}
\usage{
mriSuperResolution(image, outputDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{image}{magnetic resonance image}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
super-resolution image.
}
\description{
Perform super-resolution (2x) of MRI data using deep back projection network.
}
\examples{
\dontrun{
library( ANTsRNet )

image <- antsImageRead( "t1.nii.gz" )
imageSr <- mriSuperResolution( image )
}
}
\author{
Avants BB
}
