% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{multilabel_dice_coefficient}
\alias{multilabel_dice_coefficient}
\title{Dice function for multilabel problems}
\usage{
multilabel_dice_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
Dice value
}
\description{
Dice function for multilabel problems
}
\author{
Tustison NJ
}
