% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBrainerBrainExtraction.R
\name{noBrainerBrainExtraction}
\alias{noBrainerBrainExtraction}
\title{noBrainerBrainExtraction}
\usage{
noBrainerBrainExtraction(image, outputDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{image}{input 3-D T1-weighted brain image.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
brain mask (ANTsR image)
}
\description{
Perform "NoBrainer" brain extraction using U-net and FreeSurfer
training data.  Training and weights ported from
}
\details{
https://github.com/neuronets/nobrainer-models
}
\examples{
\dontrun{
library( ANTsRNet )
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
imageN4 <- n4BiasFieldCorrection( image )
brainMask <- noBrainerBrainExtraction( imageN4, verbose = TRUE )
}
}
\author{
Tustison NJ
}
