% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customConvolutionLayers.R
\docType{class}
\name{PartialConv2DLayer}
\alias{PartialConv2DLayer}
\title{Creates 2D partial convolution layer}
\value{
a partial convolution layer 2D
}
\description{
Creates 2D partial convolution layer as described in the paper
}
\details{
\url{https://arxiv.org/abs/1804.07723}

with the implementation ported from the following python implementation

\url{https://github.com/MathiasGruber/PConv-Keras}
}
\section{Arguments}{

}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
