% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{peak_signal_to_noise_ratio}
\alias{peak_signal_to_noise_ratio}
\title{Function to calculate peak-signal-to-noise ratio.}
\usage{
peak_signal_to_noise_ratio(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
PSNR value
}
\description{
Function to calculate peak-signal-to-noise ratio.
}
\examples{

library( ANTsRNet )
library( keras )

model <- createUnetModel2D( c( 64, 64, 1 ) )

metric_peak_signal_to_noise_ratio <-
  custom_metric( "peak_signal_to_noise_ratio",
    peak_signal_to_noise_ratio )

model \%>\% compile( loss = loss_categorical_crossentropy,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_peak_signal_to_noise_ratio ) )

}
\author{
Tustison NJ
}
