% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{pearson_correlation_coefficient}
\alias{pearson_correlation_coefficient}
\title{Model loss function for super-resolution---pearson's correlation coefficient.}
\usage{
pearson_correlation_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\description{
Based on the code found here:
}
\details{
\url{https://github.com/rstudio/keras/issues/160}

Loss functions are to be supplied in the loss parameter of the
\code{compile()} function.
}
