% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessImage.R
\name{preprocessBrainImage}
\alias{preprocessBrainImage}
\title{Basic preprocessing pipeline for T1-weighted brain MRI}
\usage{
preprocessBrainImage(
  image,
  truncateIntensity = c(0.01, 0.99),
  doBrainExtraction = TRUE,
  doBiasCorrection = TRUE,
  doDenoising = TRUE,
  intensityMatchingType = c("regression", "histogram"),
  referenceImage = NULL,
  intensityNormalizationType = c("01", "0mean"),
  verbose = TRUE
)
}
\arguments{
\item{image}{input t1-weighted brain MRI}

\item{truncateIntensity}{2-element vector giving the low and high quantiles
for intensity truncation.}

\item{doBrainExtraction}{boolean for performing ANTsRNet brain extraction
If \code{TRUE}, returns mask.  3-D input only.}

\item{doBiasCorrection}{boolean for performing N4 bias field correction.}

\item{doDenoising}{boolean for performing NLM denoising.}

\item{intensityMatchingType}{Either regression- or histogram-based.  Only is
performed if \code{!is.null(referenceImage)}.}

\item{referenceImage}{reference image for intensity matching.}

\item{intensityNormalizationType}{Either rescale the intensities to \link{0,1}
(i.e., "01") or zero-mean, unit variance (i.e., "0mean").  If \code{NULL}
normalization is not performed.}

\item{verbose}{print progress to the screen.}
}
\value{
preprocessed image and (optionally) the brain mask
}
\description{
Various preprocessing steps that have been previously described in
various papers including the cortical thickness pipeline:
}
\details{
\preformatted{    \url{https://www.ncbi.nlm.nih.gov/pubmed/24879923}
}
}
\examples{

library( ANTsRNet )

image <- antsImageRead( getANTsRData( "r16" ) )
preprocessedImage <- preprocessBrainImage( image, 
  truncateIntensity = c( 0.01, 0.99 ), doBrainExtraction = FALSE,
  doBiasCorrection = TRUE, doDenoising = TRUE, 
  intensityNormalizationType = "01", verbose = TRUE )

}
\author{
Tustison NJ, Avants BB
}
