% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{PSNR}
\alias{PSNR}
\title{Peak signal-to-noise ratio between two images.}
\usage{
PSNR(x, y)
}
\arguments{
\item{x}{input image.}

\item{y}{input image.}
}
\value{
the peak signal-to-noise ratio
}
\description{
Peak signal-to-noise ratio between two images.
}
\examples{

library( ANTsR )

r16 <- antsImageRead( getANTsRData( 'r16' ) )
r85 <- antsImageRead( getANTsRData( 'r85' ) )
psnrValue <- PSNR( r16, r85 )

}
\author{
Avants BB
}
