% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomlyTransformImageData.R
\name{randomlyHistogramWarpImageIntensity}
\alias{randomlyHistogramWarpImageIntensity}
\title{Randomly transform image intensities.}
\usage{
randomlyHistogramWarpImageIntensity(
  image,
  numberOfHistogramPoints = 4,
  sdIntensityTransform = 0.05,
  transformDomainSize = 20
)
}
\arguments{
\item{image}{input image.}

\item{numberOfHistogramPoints}{number of histogram points for intensity warping.
If = 0, no intensity warping is performed.}

\item{sdIntensityTransform}{characterize the randomness of the intensity displacement.}

\item{transformDomainSize}{Defines the sampling resolution of the B-spline warping.}
}
\value{
warped intensity image
}
\description{
Apply B-spline 1-D maps to an input image for intensity warping.
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
transformedImage <- randomlyHistogramWarpImageIntensity( image, transformDomainSize = 10 )
rm(image); gc()
rm(transformedImage); gc()
}
\author{
Tustison NJ
}
