% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomlyTransformImageData.R
\name{randomlyTransformImageData}
\alias{randomlyTransformImageData}
\title{Randomly transform image data (optional: with corresponding segmentations).}
\usage{
randomlyTransformImageData(referenceImage, inputImageList,
  segmentationImageList = NA, numberOfSimulations = 10,
  transformType = "affine", inputImageInterpolator = "linear",
  segmentationImageInterpolator = "genericLabel",
  numberOfControlPoints = 100, sdAffine = 1, spatialSmoothing = 3)
}
\arguments{
\item{referenceImage}{defines the spatial domain for all output images.  If
the input images do not match the spatial domain of the reference image, we
internally resample the target to the reference image.  This could have
unexpected consequences.  Resampling to the reference domain is performed by
testing using \code{antsImagePhysicalSpaceConsistency} then calling
\code{resampleImageToTarget} upon failure.}

\item{inputImageList}{list of lists of input images to warp.  The internal
list sets contains one or more images (per subject) which are assumed to be
mutually aligned.  The outer list contains multiple subject lists which are
randomly sampled to produce output image list.}

\item{segmentationImageList}{list of segmentation images corresponding to the
input image list (optional).}

\item{numberOfSimulations}{number of output images.  Default = 10.}

\item{transformType}{one of the following options
\code{c( "translation", "rigid", "scaleShear", "affine"," deformation" ,
  "affineAndDeformation" )}.  Default = "affine".}

\item{inputImageInterpolator}{one of the following options
\code{ c( "linear", "gaussian", "bspline" )}.  Default = "linear".}

\item{segmentationImageInterpolator}{one of the following options
\code{ c( "nearestNeighbor", "genericLabel" )}.  Default =
"nearestNeighbor".}

\item{numberOfControlPoints}{number of control points for simulated
deformations. Default = 100.}

\item{sdAffine}{parameter dictating deviation amount from identity for
random linear transformations.  Default = 1.0.}

\item{spatialSmoothing}{parameter dictating amount of spatial
smoothing.  Default = 3.0.}
}
\value{
list (if no directory set) or boolean for success, failure
}
\description{
Apply rigid, affine and/or deformable maps to an input set of training
images.  The reference image domain defines the space in which this happens.
}
\examples{

library( ANTsR )
image1 <- antsImageRead( getANTsRData( "r16" ) )
image2 <- antsImageRead( getANTsRData( "r64" ) )
segmentation1 <- thresholdImage( i1, "Otsu", 3 )
segmentation2 <- thresholdImage( i2, "Otsu", 3 )
data <- randomlyTransformImageData( image1,
  list( list( image1 ), list( image2 ) ),  
  list( segmentation1, segmentation2 ) )
}
\seealso{
\code{\link{randomImageTransformBatchGenerator}}
}
\author{
Avants BB
}
