% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{reconstructImageFromPatches}
\alias{reconstructImageFromPatches}
\title{Reconstruct image from a list of patches.}
\usage{
reconstructImageFromPatches(patchList, domainImage, strideLength = 1)
}
\arguments{
\item{patchList}{list of overlapping patches defining an image.}

\item{domainImage}{Image to define the geometric information of the
reconstructed image.}

\item{strideLength}{Defines the sequential patch overlap for
maxNumberOfPatches = all.  Can be a image-dimensional vector or a scalar.}
}
\value{
an ANTs image.
}
\description{
Reconstruct image from a list of patches.
}
\examples{
\dontrun{
}
}
\author{
Tustison NJ
}
