% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtilities.R
\name{resampleTensorLike}
\alias{resampleTensorLike}
\title{Resamples a tensor.}
\usage{
resampleTensorLike(inputTensor, referenceTensor,
  interpolationType = "nearestNeighbor")
}
\arguments{
\item{referenceTensor}{Reference tensor of rank 4 or 5 (for 2-D or 3-D volumes,
respectively).}

\item{interpolationType}{type of interpolation for resampling.  Can be
\code{nearestNeighbor}, \code{linear}, or \code{cubic}.}
}
\value{
a tensor
}
\description{
Resamples a tensor based on the reference tensor and interpolation type.
}
\examples{

library( keras )

K <- keras::backend()

inputTensor <- K$ones( c( 2L, 10L, 10L, 10L, 3L ) )
referenceTensor <- K$ones( c( 2L, 12L, 13L, 14L, 3L ) )

outputTensor <- resampleTensorLike( inputTensor, referenceTensor )

}
\author{
Tustison NJ
}
