% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTensorUtilities.R
\docType{class}
\name{ResampleTensorToTargetTensorLayer2D}
\alias{ResampleTensorToTargetTensorLayer2D}
\title{Creates a resampled tensor (to target tensor) layer (2-D)}
\value{
a resampled version of the input tensor
}
\description{
Creates a lambda layer which interpolates/resizes an input tensor based
on the specified target tensor
}
\section{Arguments}{

\describe{
\item{targetTensor}{tensor of desired size.}
\item{interpolationType}{Type of interpolation.  Can be
\code{'nearestNeighbor'}, \code{'nearestNeighbor'},
\code{'linear'}, \code{'bilinear'},
\code{'cubic'}, or \code{'bicubic'}}
\item{x}{}
\item{mask}{}
\item{input_shape}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
