% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscUtilities.R
\name{ResizeTensorLayer}
\alias{ResizeTensorLayer}
\title{Creates an interpolating (or resizing) lambda layer}
\usage{
ResizeTensorLayer(shape, interpolationType = "nearestNeighbor")
}
\arguments{
\item{shape}{vector or list of length 2 or 3 specifying the shape of the
output tensor}

\item{interpolationType}{type of interpolation for resampling.  Can be
\code{nearestNeighbor}, \code{linear}, or \code{cubic}.}
}
\value{
a keras lambda layer
}
\description{
Creates a lambda layer which interpolates/resizes an input tensor based
on the specified shape
}
\examples{

library( keras )

K <- keras::backend()

inputTensor <- K$ones( c( 10L, 10L, 3L ) )

interpolatorLayer <- ResizeTensorLayer( c( 15L, 15L ) )
outputTensor <- inputTensor \%>\% interpolatorLayer

}
\author{
Tustison NJ
}
