% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{sampleFromOutput}
\alias{sampleFromOutput}
\title{Sample from a distribution}
\usage{
sampleFromOutput(
  parameters,
  outputDimension,
  numberOfMixtures,
  temperature = 1,
  sigmaTemperature = 1
)
}
\arguments{
\item{parameters}{vector parameter to split}

\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of mixture components}

\item{temperature}{the temperature for to adjust the distribution
(default 1.0), passed to \code{\link{mixture_density_network_softmax}}}

\item{sigmaTemperature}{multipler to \code{sigma} from the
output of \code{\link{sampleFromCategoricalDistribution}}}
}
\value{
a single sample
}
\description{
Ported from:
}
\details{
\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
