% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denseUnetUtilities.R
\docType{class}
\name{ScaleLayer}
\alias{ScaleLayer}
\title{Custom scale layer}
\value{
A keras layer
}
\description{
Custom scale layer
}
\section{Arguments}{

\describe{
\item{axis}{integer specifying which axis to normalize.}
\item{momentum}{momentum value used for computation of the exponential
average of the mean and standard deviation.}
\item{input_shape}{Dimensionality of the input (integer) not
including the samples axis.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\examples{
library(keras)
inputImageSize = c( 256L, 256L, 1L )
inputs <- keras::layer_input( shape = inputImageSize )
outputs <- inputs \%>\% layer_zero_padding_2d( padding = c( 3L, 3L ) )
layer_scale = ANTsRNet:::layer_scale
outputs = outputs \%>\%
  layer_scale(axis = -1L)
lay = ScaleLayer$new()
\dontrun{
lay$build(input_shape = inputImageSize)
}
}
\author{
Tustison NJ
}
