% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\name{setANTsXNetCacheDirectory}
\alias{setANTsXNetCacheDirectory}
\title{Set the ANTsXNet Cache Directory}
\usage{
setANTsXNetCacheDirectory(path)
}
\arguments{
\item{path}{A character string specifying the path to the cache directory.}
}
\description{
This function sets the cache directory used by ANTsXNet to store downloaded
templates and model weights. If the specified directory does not exist,
it will be created.
}
