% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{shivaWmhSegmentation}
\alias{shivaWmhSegmentation}
\title{SHIVA WMH segmentation.}
\usage{
shivaWmhSegmentation(
  flair,
  t1 = NULL,
  whichModel = "all",
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{(Optional) input 3-D T1-weighted brain image (aligned to FLAIR image).}

\item{whichModel}{integer or string. Several models were trained for the
case of T1-only or T1/FLAIR image pairs.  One can use a specific single
trained model or the average of the entire ensemble.  I.e., options are:
* For T1-only:  0, 1, 2, 3, 4, 5.
* For T1/FLAIR: 0, 1, 2, 3, 4.
* Or "all" for using the entire ensemble.}

\item{doPreprocessing}{perform n4 bias correction, intensity truncation, brain
extraction.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
probabilistic image.
}
\description{
Perform segmentation of white matter hyperintensities.
\url{https://pubmed.ncbi.nlm.nih.gov/38050769/}
with the original implementation available here:
https://github.com/pboutinaud/SHIVA_WMH
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

t1 <- antsImageRead( "t1.nii.gz" )
flair <- antsImageRead( "flair.nii.gz" )
results <- wmhSegmentation( t1, flair )
}
}
\author{
Tustison NJ
}
