% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denseUnetUtilities.R,
%   R/spatialTransformerNetworkUtilities.R
\docType{class}
\name{SpatialTransformerLayer2D}
\alias{SpatialTransformerLayer2D}
\title{Custom scale layer}
\value{
.

resampled batch images.
}
\description{
Custom scale layer

Spatial transformer layer (2-D)
}
\section{Usage}{

\preformatted{outputs <- layer_scale( inputs, resampledSize )}


\preformatted{outputs <- layer_spatial_transformer_2d( inputs, resampledSize )}
}

\section{Arguments}{

\describe{
\item{axis}{integer specifying which axis to normalize.}
\item{momentum}{momentum value used for computation of the exponential
average of the mean and standard deviation.}
\item{weights}{initialization weights with shapes.}
\item{betaInitializer}{initial bias parameter. Only used if weights
aren't specified.}
\item{gammaInitializer}{initial scale parameter. Only used if weights
aren't specified.}
}


\describe{
\item{inputs}{list of size 2 where the first element are the images. The second
element are the weights.}
\item{resampledSize}{size of the resampled output images.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.


\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ

Tustison NJ
}
