% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransformerNetworkUtilities.R
\docType{class}
\name{SpatialTransformerLayer3D}
\alias{SpatialTransformerLayer3D}
\title{Spatial transfomer layer (3-D)}
\value{
resampled batch images.
}
\description{
Spatial transfomer layer (3-D)
}
\section{Usage}{

\preformatted{outputs <- layer_spatial_transformer_3d( inputs, resampledSize )}
}

\section{Arguments}{

\describe{
\item{inputs}{list of size 2 where the first element are the images. The second
element are the weights.}
\item{resampledSize}{size of the resampled output images.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
