% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{splitMixtureParameters}
\alias{splitMixtureParameters}
\title{Splits the mixture parameters.}
\usage{
splitMixtureParameters(parameters, outputDimension, numberOfMixtures)
}
\arguments{
\item{parameters}{vector parameter to split}

\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of mixture components}
}
\value{
separate mixture parameters
}
\description{
Ported from:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}\if{html}{\out{</div>}}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
