% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{ssim}
\alias{ssim}
\title{Structural similarity index}
\usage{
ssim(img1, img2, ks = c(0.01, 0.03))
}
\arguments{
\item{img1}{is a \code{antsImage} object to compare}

\item{img2}{is a \code{antsImage} object to compare}

\item{ks}{is a vector of length 2 which contains values for constants in the SSIM formula. If ignored default values will be used.}
}
\value{
\code{ssim} returns the SSIM value
}
\description{
This function implements the structural similarity index of Wang et al. (2003) formulated
for comparison of \code{antsImage} objects.
}
