% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSuperResolutionGanModel.R
\docType{class}
\name{SuperResolutionGanModel}
\alias{SuperResolutionGanModel}
\title{Super resolution GAN model}
\description{
Super resolution generative adverserial network from the paper:
}
\details{
https://arxiv.org/abs/1609.04802

and ported from the Keras (python) implementation:

https://github.com/eriklindernoren/Keras-GAN/blob/master/srgan/srgan.py
}
\section{Usage}{

}

\section{Arguments}{

\describe{
\item{inputImageSize}{}
\item{numberOfResidualBlocks}{}
}
}

\section{Details}{

\code{$initialize} {instantiates a new class and builds the
generator and discriminator.}
\code{$buildGenerator}{build generator.}
\code{$buildGenerator}{build discriminator.}
}

\examples{
\dontrun{

library( keras )
library( ANTsRNet )

keras::backend()$clear_session()

ganModel <- SuperResolutionGanModel$new(
   inputImageSize = inputImageSize )
}

}
\author{
Tustison NJ
}
