% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{sysuMediaWmhSegmentation}
\alias{sysuMediaWmhSegmentation}
\title{White matter hyperintensity segmentation}
\usage{
sysuMediaWmhSegmentation(flair, t1 = NULL, useEnsemble = TRUE, verbose = FALSE)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{input 3-D T1-weighted brain image (assumed to be aligned to
the flair, if specified).}

\item{useEnsemble}{boolean to check whether to use all 3 sets of weights.}

\item{verbose}{print progress.}
}
\value{
WMH segmentation probability image
}
\description{
Perform WMH segmentation using the winning submission in the MICCAI
2017 challenge by the sysu_media team using FLAIR or T1/FLAIR.  The
MICCAI challenge is discussed in
}
\details{
\url{https://pubmed.ncbi.nlm.nih.gov/30908194/}

with the sysu_media's team entry is discussed in

\url{https://pubmed.ncbi.nlm.nih.gov/30125711/}

with the original implementation available here:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\url{https://github.com/hongweilibran/wmh_ibbmTum}
}\if{html}{\out{</div>}}

The original implementation used global thresholding as a quick
brain extraction approach.  Due to possible generalization difficulties,
we leave such post-processing steps to the user.  For brain or white
matter masking see functions brainExtraction or deepAtropos,
respectively.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "flair.nii.gz" )
probabilityMask <-sysuMediaWmhSegmentation( image )
}
}
\author{
Tustison NJ
}
