% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityAssessment.R
\name{tidNeuralImageAssessment}
\alias{tidNeuralImageAssessment}
\title{Perform TID-based assessment of an image.}
\usage{
tidNeuralImageAssessment(
  image,
  mask,
  patchSize = 101L,
  strideLength = 7L,
  paddingSize = 0L,
  dimensionsToPredict = 1,
  outputDirectory = NULL,
  whichModel = "tidsQualityAssessment",
  verbose = FALSE
)
}
\arguments{
\item{image}{the input image.  Either 2D or 3D.}

\item{mask}{optional mask for designating calculation ROI.}

\item{patchSize}{integer prime number for patch size; 101 is good. otherwise,
choose \code{"global"} for a single global estimate of quality.}

\item{strideLength}{optional value to speed up computation (typically less than
patch size).  Integer or vector of image dimension length.}

\item{paddingSize}{positive or negative integer (or vector of image dimension
length) for (de)padding to remove edge effects.}

\item{dimensionsToPredict}{if image dimension is 3, this parameter specifies
which dimension(s) should be used for prediction.  If more than one dimension
is specified, the results are averaged.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{whichModel}{model type e.g. string tidsQualityAssessment, koniqMBCS}

\item{verbose}{print progress.}
}
\value{
list of QC results predicting both both human rater's mean and standard
deviation of the MOS ("mean opinion scores").  See the TID2013 paper.  Both
aggregate and spatial scores are returned, the latter in the form of an image.
}
\description{
Use a ResNet architecture to estimate image quality in 2D or 3D using the TID
image database described in
}
\details{
\url{https://www.sciencedirect.com/science/article/pii/S0923596514001490}

where the image assessment is either "global", i.e., a single number or an image
based on the specified patch size.  In the 3-D case, neighboring slices are used
for each estimate.  Note that parameters should be kept as consistent as possible
in order to enable comparison.  Patch size should be roughly 1/12th to 1/4th of
image size to enable locality. A global estimate can be gained by setting
\code{patchSize = "global"}.
}
\examples{
\dontrun{
image <- antsImageRead( getANTsRData( "r16" ) )
mask <- getMask( image )
tid <- tidQualityAssessment( image, mask = mask, patchSize = 101L,
          strideLength = 7L, paddingSize = 0L )
plot( image, tid$MOS, alpha = 0.5)
cat( "mean MOS = ", tid$MOS.mean, "\n" )
cat( "sd MOS = ", tid$MOS.standardDeviationMean, "\n" )
}
}
\author{
Avants BB
}
