% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvaSeg.R
\name{uvaSeg}
\alias{uvaSeg}
\title{Unsupervised variational autoencoder segmentation}
\usage{
uvaSeg(
  image,
  model,
  k,
  mask,
  returnProbabilities = FALSE,
  batchSize = 1028,
  standardize = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{image}{input image}

\item{model}{the model output from \code{uvaSegTrain}}

\item{k}{number of clusters or cluster centers}

\item{mask}{defining output segmentation space}

\item{returnProbabilities}{boolean}

\item{batchSize}{for the prediction}

\item{standardize}{boolean controlling whether patches are standardized}

\item{verbose}{boolean}
}
\value{
segmentation and probability images are output
}
\description{
Trains a variational autoencoding with a convolutional network. This is
followed by k-means clustering to produce a segmentation and probabilities.
}
\examples{
\donttest{
library(ANTsR)
img <- ri( 1 )  \%>\% resampleImage( c(4,4) )  \%>\% iMath( "Normalize" )
mask = randomMask( getMask( img ), 50 )
patch = getNeighborhoodInMask( img, mask, c(3,3), boundary.condition = "NA" )
uvaSegModel = uvaSegTrain( patch, k = 6 )
tarImg = ri( 3 ) \%>\% \%>\% resampleImage( c(4,4) )
uvaSegmentation = uvaSeg(tarImg, uvaSegModel, k = 3, getMask( tarImg ) )
}
}
\author{
Avants BB
}
