% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{weighted_categorical_crossentropy}
\alias{weighted_categorical_crossentropy}
\title{Function for weighted categorical cross entropy}
\usage{
weighted_categorical_crossentropy(y_true, y_pred, weights)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}

\item{weights}{weights for each class}
}
\value{
function value
}
\description{
ported from this implementation:
}
\details{
https://gist.github.com/wassname/ce364fddfc8a025bfab4348cf5de852d
}
\examples{

library( ANTsRNet )
library( keras )

model <- createUnetModel2D( c( 64, 64, 1 ), numberOfOutputs = 2 )

model \%>\% compile( loss = weighted_categorical_crossentropy( weights = c( 1, 1 ) ),
 optimizer = optimizer_adam( lr = 0.0001 ),
   metrics = "accuracy" )

}
\author{
Tustison NJ
}
