% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpainting.R
\name{wholeHeadInpainting}
\alias{wholeHeadInpainting}
\title{wholeHeadInpainting}
\usage{
wholeHeadInpainting(
  image,
  roiMask,
  modality = "t1",
  slicewise = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D MR image.}

\item{roiMask}{binary mask image}

\item{modality}{Modality image type.  Options include: "t1": T1-weighted MRI.
"flair": FLAIR MRI.}

\item{slicewise}{Two models per modality are available for processing the data.  One model
is based on training/prediction using 2-D axial slice data whereas the
other uses 64x64x64 patches.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
inpainted image
}
\description{
Perform in-painting for whole-head MRI
}
\author{
Tustison NJ
}
