% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{wmhSegmentation}
\alias{wmhSegmentation}
\title{White matter hyperintensity probabilistic segmentation}
\usage{
wmhSegmentation(
  flair,
  t1,
  whiteMatterMask = NULL,
  useCombinedModel = TRUE,
  predictionBatchSize = 16,
  patchStrideLength = 32,
  doPreprocessing = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{input 3-D T1-weighted brain image (assumed to be aligned to
the flair).}

\item{whiteMatterMask}{input white matter mask for patch extraction.  If None,
calculated using deepAtropos (labels 3 and 4).}

\item{useCombinedModel}{Original or combined.}

\item{predictionBatchSize}{Control memory usage for prediction.  More consequential
for GPU-usage.}

\item{patchStrideLength}{3-D vector or int.   Dictates the stride length for
accumulating predicting patches.}

\item{doPreprocessing}{perform n4 bias correction, intensity truncation, brain
extraction.}

\item{verbose}{print progress.}
}
\value{
probabilistic image.
}
\description{
Perform white matter hyperintensity probabilistic segmentation
using deep learning
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

t1 <- antsImageRead( "t1.nii.gz" )
flair <- antsImageRead( "flair.nii.gz" )
results <- wmhSegmentation( t1, flair )
}
}
\author{
Tustison NJ
}
