% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{class}
\name{AnchorBoxLayer2D}
\alias{AnchorBoxLayer2D}
\title{Anchor box layer for SSD architecture (2-D).}
\value{
a 5-D tensor with shape
\eqn{ batchSize \times widthSize \times heightSize \times numberOfBoxes \times 8 }
In the last dimension, the first 4 values correspond to the
2-D coordinates of the bounding boxes and the other 4 are the variances.
}
\description{
Anchor box layer for SSD architecture (2-D).
}
\section{Usage}{

\preformatted{anchorBoxGenerator <- AnchorBoxLayer2D$new( imageSize,
     scale, nextScale, aspectRatios = c( '1:1', '2:1', '1:2' ),
     variances = 1.0 )

anchorBoxGenerator$call( x, mask = NULL )
anchorBoxGenerator$compute_output_shape( input_shape )
}
}

\section{Arguments}{

\describe{
\item{anchorBoxGenerator}{A \code{process} object.}
\item{imageSize}{size of the input image.}
\item{scale}{scale of each box (in pixels).}
\item{nextScale}{next scale of each box (in pixels).}
\item{aspectRatios}{vector describing the geometries of the anchor boxes
for this layer.}
\item{variances}{a list of 4 floats > 0 with scaling factors for the encoded
predicted box coordinates. A variance value of 1.0 would apply no scaling at
all to the predictions, while values in (0,1) upscale the encoded
predictions and values greater than 1.0 downscale the encoded predictions.
Defaults to 1.0.}
\item{x}{}
\item{mask}{}
\item{input_shape}{}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\examples{
x = AnchorBoxLayer2D$new(imageSize = c(20, 20),
scale = 2, nextScale = 2)
x$build()
}
\author{
Tustison NJ
}
