% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{data}
\name{AnchorBoxLayer3D}
\alias{AnchorBoxLayer3D}
\title{Anchor box layer for SSD architecture (3-D).}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnchorBoxLayer3D
}
\arguments{
\item{imageSize}{size of the input image.}

\item{scale}{scale of each box (in pixels).}

\item{nextScale}{next scale of each box (in pixels).}

\item{aspectRatios}{vector describing the geometries of the anchor
boxes for this layer.}

\item{variances}{explained here:
           https://github.com/rykov8/ssd_keras/issues/53

Input shape:}
}
\value{
a 6-D tensor with shape
    [batchSize, widthSize, heightSize, depthSize, numberOfBoxes, 12]
In the last dimension, the first 6 values correspond to the
3-D coordinates of the bounding boxes and the other 6 are the variances.
}
\description{
Anchor (or prior) boxes for SSD architecture (3-D).  The input shape is
defined as
}
\details{
Theano:  [batchSize, channelSize, widthSize, heightSize, depthSize]
    tensorflow:  [batchSize, widthSize, heightSize, depthSize, channelSize]
}
\examples{

\dontrun{ 

library( keras )

}
}
\author{
Tustison NJ
}
\keyword{datasets}
