% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDBPN4x.R
\name{applyDBPNsr}
\alias{applyDBPNsr}
\title{applyDBPNsr}
\usage{
applyDBPNsr(image, modelFilename, expansionFactor, verbose = FALSE)
}
\arguments{
\item{image}{input image}

\item{modelFilename}{see \code{getPretrainedNetwork}}

\item{expansionFactor}{size to upscale, should match network pretraining}

\item{verbose}{If \code{TRUE}, show status messages}
}
\value{
image upscaled to resolution provided by network
}
\description{
Apply pretrained DBPN super-resolution network
}
\examples{
\dontrun{
simg <- applyDBPNsr( ri( 1 ), getPretrainedNetwork( "dbpn4x" ) )
}
}
\author{
Avants BB
}
