% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histology.R
\name{arterialLesionSegmentation}
\alias{arterialLesionSegmentation}
\title{Arterial lesion segmentation}
\usage{
arterialLesionSegmentation(
  image,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D lung image.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Perform arterial lesion segmentation using U-net.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "image.nii.gz" )
output <- lungExtraction( image )
}
}
\author{
Tustison NJ
}
