% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainAge.R
\name{brainAge}
\alias{brainAge}
\title{BrainAGE}
\usage{
brainAge(image, doPreprocessing = TRUE, outputDirectory = NULL, verbose = TRUE)
}
\arguments{
\item{image}{input 3-D T1-weighted brain image.}

\item{doPreprocessing}{boolean dictating whether prescribed
preprocessing is performed (brain extraction, bias correction,
normalization to template).}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
predicted age and binned confidence values
}
\description{
Estimate BrainAge from a T1-weighted MR image using the DeepBrainNet
architecture and weights described here:
}
\details{
\url{https://github.com/vishnubashyam/DeepBrainNet}

and described in the following article:

\url{https://academic.oup.com/brain/article-abstract/doi/10.1093/brain/awaa160/5863667?redirectedFrom=fulltext}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item brain extraction, and
\item affine registration to MNI.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
estimatedBrainAge <- brainAge( image )
}
}
\author{
Tustison NJ
}
