% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainExtraction.R
\name{brainExtraction}
\alias{brainExtraction}
\title{Brain extraction}
\usage{
brainExtraction(
  image,
  modality = c("t1", "bold", "fa", "t1nobrainer"),
  outputDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D brain image.}

\item{modality}{image type.  Options include "t1", "bold", "fa", "t1nobrainer".}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
brain probability mask (ANTsR image)
}
\description{
Perform T1, FA, or bold brain extraction using a U-net architecture
training data.  "NoBrainer" is also possible where
brain extraction uses U-net and FreeSurfer
training data ported from the
}
\details{
https://github.com/neuronets/nobrainer-models
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1w_image.nii.gz" )
probabilityMask <- brainExtraction( image, modality = "t1" )
}
}
\author{
Tustison NJ
}
