% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chexnet.R
\name{checkXrayLungOrientation}
\alias{checkXrayLungOrientation}
\title{Check x-ray lung orientation.}
\usage{
checkXrayLungOrientation(image, antsxnetCacheDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{image}{input 3-D lung image.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Check the correctness of image orientation, i.e., flipped left-right, up-down,
or both.  If True, attempts to correct before returning corrected image.  Otherwise
it returns NULL.
}
\examples{
\dontrun{
library( ANTsRNet )
}
}
\author{
Tustison NJ
}
