% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chexnet.R
\name{chexnet}
\alias{chexnet}
\title{CheXNet: Radiologist-Level Pneumonia Detection on Chest X-Rays with Deep Learning}
\usage{
chexnet(
  image,
  lungMask = NULL,
  checkImageOrientation = FALSE,
  useANTsXNetVariant = TRUE,
  includeTuberculosisDiagnosis = FALSE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D lung image.}

\item{lungMask}{ANTsImage.  If None is specified, one is estimated.}

\item{checkImageOrientation}{Check the correctness of image orientation, i.e.,
flipped left-right, up-down, or both.  If TRUE, attempts to correct before prediction.}

\item{useANTsXNetVariant}{Use an extension of the original chexnet approach
by adding a left/right lung masking and including those two masked regions
in the red and green channels, respectively.}

\item{includeTuberculosisDiagnosis}{Include the output of an additional network
trained on TB data but using the ANTsXNet variant chexnet data as the initial
weights.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
classification scores for each of the 14 or 15 categories.
}
\description{
ANTsXNet reproduction of https://arxiv.org/pdf/1711.05225.pdf.  This includes
our own network architecture and training (including data augmentation).

Includes a network for checking the correctness of image orientation, i.e.,
flipped left-right, up-down, or both.

Disease categories:
Atelectasis
Cardiomegaly
Consolidation
Edema
Effusion
Emphysema
Fibrosis
Hernia
Infiltration
Mass
No Finding
Nodule
Pleural Thickening
Pneumonia
Pneumothorax
}
\examples{
\dontrun{
library( ANTsRNet )
}
}
\author{
Tustison NJ
}
