% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claustrumSegmentation.R
\name{claustrumSegmentation}
\alias{claustrumSegmentation}
\title{Claustrum segmentation}
\usage{
claustrumSegmentation(
  t1,
  doPreprocessing = TRUE,
  useEnsemble = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{t1}{input 3-D T1-weighted brain image.}

\item{doPreprocessing}{perform n4 bias correction, denoising?}

\item{useEnsemble}{boolean to check whether to use all 3 sets of weights.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
claustrum probability image
}
\description{
Described here:
}
\details{
\preformatted{\url{https://pubmed.ncbi.nlm.nih.gov/34520080/}
}

with the implementation available at:\preformatted{\url{https://github.com/hongweilibran/claustrum_multi_view}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
probabilityMask <- claustrumSegmentation( image )
}
}
\author{
Tustison NJ
}
