% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEmbeddedClusteringUtilities.R
\name{createAutoencoderModel}
\alias{createAutoencoderModel}
\title{Function for creating a symmetric autoencoder model.}
\usage{
createAutoencoderModel(numberOfUnitsPerLayer, activation = "relu",
  initializer = "glorot_uniform")
}
\arguments{
\item{numberOfUnitsLayer}{an array of defining the number of units
in the encoding branch}
}
\value{
two models:  the encoder and auto-encoder
}
\description{
Builds an autoencoder based on the specified array definining the
number of units in the encoding branch.  Ported to Keras R from the
Keras python implementation here:
}
\details{
\url{https://github.com/XifengGuo/DEC-keras/blob/master/DEC.py}
}
\author{
Tustison NJ
}
