% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUpDownResNetSuperResolutionModel.R
\name{createEnhancedDeepSuperResolutionModel2D}
\alias{createEnhancedDeepSuperResolutionModel2D}
\title{2-D implementation of the EDSR super resolution architecture.}
\usage{
createEnhancedDeepSuperResolutionModel2D(
  inputImageSize,
  convolutionKernelSize = c(3, 3),
  numberOfFilters = 256,
  numberOfResidualBlocks = 32,
  scale = 2,
  numberOfLossFunctions = 1,
  numberOfOutputChannels = 1,
  doBatchNormalization = FALSE,
  interpolation = c("bilinear", "nearest", "conv")
)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{convolutionKernelSize}{a vector specifying the kernel size for
convolution.}

\item{numberOfFilters}{the number of filters for each encoding layer.}

\item{numberOfResidualBlocks}{the number of residual blocks.}

\item{scale}{the upsampling amount, 2, 4 or 8}

\item{numberOfLossFunctions}{the number of data targets, e.g. 2 for 2 targets}

\item{numberOfOutputChannels}{the number of ouput channels}

\item{doBatchNormalization}{boolean for include BN in the residual blocks}

\item{interpolation}{nearest, bilinear or conv for upscaling block}
}
\value{
a keras model for EDSR image super resolution
}
\description{
Creates a keras model of the expanded image super resolution deep learning
framework based on EDSR.
}
\examples{
createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ))
createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ),
doBatchNormalization = TRUE,
interpolation = "conv", scale = 4)
createEnhancedDeepSuperResolutionModel2D(c( 28, 28, 1 ),
doBatchNormalization = TRUE,
numberOfLossFunctions = 2,
interpolation = "conv", scale = 8)
}
\author{
Tustison NJ, Avants BB
}
