% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createHippMapp3rUnetModel3D}
\alias{createHippMapp3rUnetModel3D}
\title{Implementation of the "HippMapp3r" U-net architecture}
\usage{
createHippMapp3rUnetModel3D(inputImageSize, doFirstNetwork = TRUE)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
described here:
}
\details{
\url{https://onlinelibrary.wiley.com/doi/pdf/10.1002/hbm.24811}

with the implementation available here:

\url{https://github.com/mgoubran/HippMapp3r}
}
\examples{
\dontrun{

model1 <- createHippMapp3rUnetModel3D( c( 160, 160, 128, 1 ), doFirstNetwork = TRUE )
model2 <- createHippMapp3rUnetModel3D( c( 112, 112, 64, 1 ), doFirstNetwork = FALSE )

}
}
\author{
Tustison NJ
}
