% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createNoBrainerUnetModel3D}
\alias{createNoBrainerUnetModel3D}
\title{Implementation of the "NoBrainer" U-net architecture}
\usage{
createNoBrainerUnetModel3D(inputImageSize)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
avaialable here:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    \url{https://github.com/neuronets/nobrainer/}
}\if{html}{\out{</div>}}
}
\examples{

library( ANTsRNet )

model <- createNoBrainerUnetModel3D( list( NULL, NULL, NULL, 1 ) )

}
\author{
Tustison NJ
}
