% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomModel.R
\name{createRmnetGenerator}
\alias{createRmnetGenerator}
\title{Implementation of the "RMNet" generator architecture for inpainting}
\usage{
createRmnetGenerator(numberOfChannels = 3)
}
\value{
a keras model
}
\description{
Creates a keras model implementation of the model:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    \url{https://github.com/Jireh-Jam/R-MNet-Inpainting-keras}
}\if{html}{\out{</div>}}
}
\examples{

library( ANTsRNet )

model <- createRmnetGenerator()

}
\author{
Tustison NJ
}
