% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCustomUnetModel.R
\name{createSysuMediaUnetModel2D}
\alias{createSysuMediaUnetModel2D}
\title{Implementation of the sysu_media U-net architecture}
\usage{
createSysuMediaUnetModel2D(inputImageSize)
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.
This will be \code{c(200, 200, 2)} for t1/flair input and
\code{c(200, 200, 1)} for flair-only input.}
}
\value{
a u-net keras model
}
\description{
Creates a keras model implementation of the u-net architecture
in the 2017 MICCAI WMH challenge by the sysu_medial team described
here:
}
\details{
\preformatted{\url{https://pubmed.ncbi.nlm.nih.gov/30125711/}
}

with the original implementation available here:\preformatted{\url{https://github.com/hongweilibran/wmh_ibbmTum}
}
}
\examples{
\dontrun{

model <- createLiWmhUnetModel2D( c( 200, 200, 1 ) )

}
}
\author{
Tustison NJ
}
