% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVggModel.R
\name{createVggModel3D}
\alias{createVggModel3D}
\title{3-D implementation of the Vgg deep learning architecture.}
\usage{
createVggModel3D(inputImageSize, numberOfClassificationLabels = 1000,
  layers = c(1, 2, 3, 4, 4), lowestResolution = 64,
  convolutionKernelSize = c(3, 3, 3), poolSize = c(2, 2, 2),
  strides = c(2, 2, 2), denseUnits = 4096, dropoutRate = 0,
  style = 19, mode = "classification")
}
\arguments{
\item{inputImageSize}{Used for specifying the input tensor shape.  The
shape (or dimension) of that tensor is the image dimensions followed by
the number of channels (e.g., red, green, and blue).  The batch size
(i.e., number of training images) is not specified a priori.}

\item{numberOfClassificationLabels}{Number of segmentation labels.}

\item{layers}{a vector determining the number of 'filters' defined at
for each layer.}

\item{lowestResolution}{number of filters at the beginning.}

\item{convolutionKernelSize}{3-d vector definining the kernel size
during the encoding path}

\item{poolSize}{3-d vector defining the region for each pooling layer.}

\item{strides}{3-d vector describing the stride length in each direction.}

\item{denseUnits}{integer for the number of units in the last layers.}

\item{dropoutRate}{float between 0 and 1 to use between dense layers.}

\item{style}{\verb{'16'} or \verb{'19'} for VGG16 or VGG19, respectively.}

\item{mode}{'classification' or 'regression'.  Default = 'classification'.}
}
\value{
a VGG keras model to be used with subsequent fitting
}
\description{
Creates a keras model of the Vgg deep learning architecture for image
recognition based on the paper
}
\details{
K. Simonyan and A. Zisserman, Very Deep Convolutional Networks for
Large-Scale Image Recognition

available here:

\url{https://arxiv.org/abs/1409.1556}

This particular implementation was influenced by the following python
implementation:

\url{https://gist.github.com/baraldilorenzo/8d096f48a1be4a2d660d}
}
\examples{
# Simple examples, must run successfully and quickly. These will be tested.
\dontrun{

library( keras )

mnistData <- dataset_mnist()

numberOfLabels <- length( unique( mnistData$train$y ) )

X_train <- array( mnistData$train$x, dim = c( dim( mnistData$train$x ), 1 ) )
Y_train <- keras::to_categorical( mnistData$train$y, numberOfLabels )

# we add a dimension of 1 to specify the channel size
inputImageSize <- c( dim( mnistData$train$x )[2:3], 1 )

vggModel <- createVggModel2D( inputImageSize = inputImageSize,
  numberOfClassificationLabels = numberOfLabels, layers = c( 1, 4, 4 ),
  lowestResolution = 14 )

vggModel \%>\% compile( loss = 'categorical_crossentropy',
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( 'categorical_crossentropy', 'accuracy' ) )

track <- vggModel \%>\% fit( X_train, Y_train, epochs = 40, batch_size = 32,
  verbose = 1, shuffle = TRUE, validation_split = 0.2 )

# Now test the model

X_test <- array( mnistData$test$x, dim = c( dim( mnistData$test$x ), 1 ) )
Y_test <- keras::to_categorical( mnistData$test$y, numberOfLabels )

testingMetrics <- vggModel \%>\% evaluate( X_test, Y_test )
predictedData <- vggModel \%>\% predict( X_test, verbose = 1 )

}
}
\author{
Tustison NJ
}
