% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCycleGanModel.R
\docType{class}
\name{CycleGanModel}
\alias{CycleGanModel}
\title{Cycle GAN model}
\description{
Deep conviolutional generative adverserial network from the paper:
}
\details{
https://arxiv.org/pdf/1703.10593

and ported from the Keras (python) implementation:

https://github.com/eriklindernoren/Keras-GAN/blob/master/cyclegan/cyclegan.py
}
\section{Usage}{

}

\section{Arguments}{

\describe{
\item{inputImageSize}{}
\item{latentDimension}{}
}
}

\section{Details}{

\code{$initialize} {instantiates a new class and builds the
generator and discriminator.}
\code{$buildGenerator}{build generator.}
\code{$buildGenerator}{build discriminator.}
}

\examples{
\dontrun{

library( keras )
library( ANTsRNet )

keras::backend()$clear_session()

ganModel <- CycleGanModel$new(
   inputImageSize = inputImageSize )
}

}
\author{
Tustison NJ
}
