% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepAtropos.R
\name{deepAtropos}
\alias{deepAtropos}
\title{Six tissue segmentation}
\usage{
deepAtropos(
  t1,
  doPreprocessing = TRUE,
  useSpatialPriors = 0,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE,
  debug = FALSE
)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted brain image.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{useSpatialPriors}{use MNI spatial tissue priors (0, 1, or 2).  0 is no priors.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}

\item{debug}{return feature images in the last layer of the u-net model.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform Atropos-style six tissue segmentation using deep learning
}
\details{
The labeling is as follows:
\itemize{
\item{Label 0:}{background}
\item{Label 1:}{CSF}
\item{Label 2:}{gray matter}
\item{Label 3:}{white matter}
\item{Label 4:}{deep gray matter}
\item{Label 5:}{brain stem}
\item{Label 6:}{cerebellum}
}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item denoising,
\item brain extraction, and
\item affine registration to MNI.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- deepAtropos( image )
}
}
\author{
Tustison NJ
}
