% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepEmbeddedClusteringUtilities.R
\docType{class}
\name{DeepEmbeddedClusteringModel}
\alias{DeepEmbeddedClusteringModel}
\title{Deep embedded clustering (DEC) model class}
\description{
Deep embedded clustering (DEC) model class
}
\section{Usage}{

}

\section{Arguments}{

\describe{
\item{numberOfUnitsPerLayer}{array describing the auteoencoder.}
\item{numberOfClusters}{number of clusters.}
\item{alpha}{parameter}
\item{initializer}{initializer for autoencoder}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$pretrain}

\code{$loadWeights}

\code{$extractFeatures}

\code{$predictClusterLabels}

\code{$targetDistribution}

\code{$compile}

\code{$fit}
}

\author{
Tustison NJ
}
