% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepFlash.R
\name{deepFlash}
\alias{deepFlash}
\title{Hippocampal/Enthorhinal segmentation using "Deep Flash"}
\usage{
deepFlash(
  t1,
  doPreprocessing = TRUE,
  outputDirectory = NULL,
  verbose = FALSE,
  debug = FALSE
)
}
\arguments{
\item{t1}{raw or preprocessed 3-D T1-weighted brain image.}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}

\item{debug}{return feature images in the last layer of the u-net model.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform hippocampal/entorhinal segmentation in T1 images using
labels from Mike Yassa's lab
}
\details{
\url{https://faculty.sites.uci.edu/myassa/}

The labeling is as follows:
\itemize{
\item{Label 0 :}{background}
\item{Label 5 :}{left aLEC}
\item{Label 6 :}{right aLEC}
\item{Label 7 :}{left pMEC}
\item{Label 8 :}{right pMEC}
\item{Label 9 :}{left perirhinal}
\item{Label 10:}{right perirhinal}
\item{Label 11:}{left parahippocampal}
\item{Label 12:}{right parahippocampal}
\item{Label 13:}{left DG/CA3}
\item{Label 14:}{right DG/CA3}
\item{Label 15:}{left CA1}
\item{Label 16:}{right CA1}
\item{Label 17:}{left subiculum}
\item{Label 18:}{right subiculum}
}

Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item denoising,
\item brain extraction, and
\item affine registration to MNI.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "t1.nii.gz" )
results <- deepFlash( image )
}
}
\author{
Tustison NJ
}
