% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{ewDavid}
\alias{ewDavid}
\title{White matter hypterintensity probabilistic segmentation}
\usage{
ewDavid(
  flair,
  t1,
  doPreprocessing = TRUE,
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{input 3-D T1-weighted brain image (assumed to be aligned to
the flair).}

\item{doPreprocessing}{perform preprocessing.  See description above.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}

\item{debug}{return feature images in the last layer of the u-net model.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform White matter hypterintensity probabilistic segmentation
using deep learning
}
\details{
Preprocessing on the training data consisted of:
\itemize{
\item n4 bias correction,
\item brain extraction, and
\item affine registration to MNI.
The input T1 should undergo the same steps.  If the input T1 is the raw
T1, these steps can be performed by the internal preprocessing, i.e. set
\code{doPreprocessing = TRUE}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

t1 <- antsImageRead( "t1.nii.gz" )
flair <- antsImageRead( "flair.nii.gz" )
results <- ewDavid( t1, flair )
}
}
\author{
Tustison NJ
}
