% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{ewDavid}
\alias{ewDavid}
\title{White matter hypterintensity probabilistic segmentation}
\usage{
ewDavid(
  flair,
  t1,
  doPreprocessing = TRUE,
  whichModel = "sysu",
  whichAxes = "max",
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{input 3-D T1-weighted brain image (assumed to be aligned to
the flair).}

\item{whichModel}{one of "sysu", "sysuT1Only", "sysuPlus",
"sysuPlusSeg", "sysuWithSite", "sysuWithSiteT1Only".}

\item{whichAxes}{apply 2-D model to 1 or more axes.  In addition to a scalar
or vector, e.g., \code{whichAxes = c(1, 3)}, one can use "max" for the
axis with maximum anisotropy (default) or "all" for all axes.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}

\item{debug}{return feature images in the last layer of the u-net model.}
}
\value{
list consisting of the segmentation image and probability images for
each label.
}
\description{
Perform white matter hypterintensity probabilistic segmentation
using deep learning
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

t1 <- antsImageRead( "t1.nii.gz" )
flair <- antsImageRead( "flair.nii.gz" )
results <- ewDavid( t1, flair )
}
}
\author{
Tustison NJ
}
