% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{extractImagePatches}
\alias{extractImagePatches}
\title{Extract 2-D or 3-D image patches.}
\usage{
extractImagePatches(image, patchSize, maxNumberOfPatches = "all",
  strideLength = 1, randomSeed)
}
\arguments{
\item{image}{Input ANTs image}

\item{patchSize}{Width, height, and depth (if 3-D) of patches.}

\item{maxNumberOfPatches}{Maximum number of patches returned.  If
"all" is specified, then all patches in sequence (defined by the}

\item{strideLength}{Defines the sequential patch overlap for
maxNumberOfPatches = all.  Can be a image-dimensional vector or a scalar.}

\item{randomSeed}{integer seed that allows reproducible patch extraction
across runs.}
}
\value{
a randomly selected list of patches.
}
\description{
Extract 2-D or 3-D image patches.
}
\examples{

library( ANTsR )
image <- antsImageRead( getANTsRData( "r16" ) )
patchSet1 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 0 )
patchSet2 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 1 )
patchSet3 <- extractImagePatches( image, c( 32, 32 ), 10, c( 32, 32 ), randomSeed = 0 )

}
\author{
Tustison NJ
}
