% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{extractImagePatches}
\alias{extractImagePatches}
\title{Extract 2-D or 3-D image patches.}
\usage{
extractImagePatches(image, patchSize, maxNumberOfPatches = "all")
}
\arguments{
\item{image}{Input ANTs image}

\item{patchSize}{Width, height, and depth (if 3-D) of patches.}

\item{maxNumberOfPatches}{Maximum number of patches returned.  If
"all" is specified, then all overlapping patches are extracted.}
}
\value{
a randomly selected list of patches.
}
\description{
Extract 2-D or 3-D image patches.
}
\examples{
\dontrun{ 
}
}
\author{
Tustison NJ
}
