% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\name{getANTsXNetCacheDirectory}
\alias{getANTsXNetCacheDirectory}
\title{Get the ANTsXNet Cache Directory}
\usage{
getANTsXNetCacheDirectory()
}
\value{
A character string specifying the path to the cache directory.
}
\description{
This function returns the current cache directory used by ANTsXNet to store
downloaded templates and model weights. If no directory has been set, it
initializes it to the default path "~/.keras/ANTsXNet".
}
