% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getANTsXNetData.R
\name{getANTsXNetData}
\alias{getANTsXNetData}
\title{getANTsXNetData}
\usage{
getANTsXNetData(
  fileId = c("show", "biobank", "croppedMni152", "mprage_hippmapp3r",
    "protonLungTemplate", "ctLungTemplate", "priorDktLabels", "priorDeepFlashLeftLabels",
    "priorDeepFlashRightLabels", "S_template3", "adni", "ixi", "kirby", "mni152", "nki",
    "nki10", "oasis"),
  targetFileName,
  antsxnetCacheDirectory = NULL
)
}
\arguments{
\item{fileId}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{targetFileName}{optional target filename}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}
}
\value{
filename string
}
\description{
Download data such as prefabricated templates and spatial priors.
}
\note{
See \url{https://figshare.com/authors/Nick_Tustison/441144}
or \url{https://figshare.com/authors/Brian_Avants/418551}
for some more descriptions
}
\examples{
\dontrun{
net <- getANTsXNetData("biobank")
}
}
\author{
Avants BB
}
