% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{getMixtureDensitySamplingFunction}
\alias{getMixtureDensitySamplingFunction}
\title{Returns a sampling function for the mixture density.}
\usage{
getMixtureDensitySamplingFunction(outputDimension, numberOfMixtures)
}
\arguments{
\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of mixture components}
}
\value{
a function for sampling a mixture density
}
\description{
Ported from:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}\if{html}{\out{</div>}}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
