% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{getMixtureDensitySamplingFunction}
\alias{getMixtureDensitySamplingFunction}
\title{Returns a sampling function for the mixture density.}
\usage{
getMixtureDensitySamplingFunction(outputDimension, numberOfMixes)
}
\arguments{
\item{outputDimension}{output dimension}

\item{numberOfMixes}{number of mixture components}
}
\value{
a function for sampling a mixture density
}
\description{
Ported from:
}
\details{
https://github.com/cpmpercussion/keras-mdn-layer/
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
