% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPretrainedNetwork.R
\name{getPretrainedNetwork}
\alias{getPretrainedNetwork}
\title{getPretrainedNetwork}
\usage{
getPretrainedNetwork(fileId, targetFileName)
}
\arguments{
\item{targetFileName}{optional target filename}

\item{fileid}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}
}
\value{
filename string
}
\description{
Downloads antsrnet pretrained network. \url{10.6084/m9.figshare.7246985}
}
\examples{
\dontrun{
net <- getPretrainedNetwork( "dbpn4x" )
}
}
\author{
Avants BB
}
