% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPretrainedNetwork.R
\name{getPretrainedNetwork}
\alias{getPretrainedNetwork}
\title{getPretrainedNetwork}
\usage{
getPretrainedNetwork(
  fileId = c("show", "brainAgeGender", "brainAgeFmrib", "brainExtraction",
    "brainSegmentation", "brainSegmentationPatchBased", "ctHumanLung", "dbpn4x",
    "denoising", "functionalLungMri", "hippMapp3rInitial", "hippMapp3rRefine",
    "mriSuperResolution", "protonLungMri", "wholeTumorSegmentationT2Flair"),
  targetFileName,
  overwrite = FALSE
)
}
\arguments{
\item{fileId}{one of the permitted file ids or pass "show" to list all
valid possibilities. Note that most require internet access to download.}

\item{targetFileName}{optional target filename}

\item{overwrite}{shoudl the file be overwritten}
}
\value{
filename string
}
\description{
Downloads pretrained network/weights.
}
\note{
See \url{https://figshare.com/authors/Nick_Tustison/441144}
or \url{https://figshare.com/authors/Brian_Avants/418551}
for some more descriptions
}
\examples{
\dontrun{
net <- getPretrainedNetwork( "dbpn4x" )
}
}
\author{
Avants BB
}
