% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hippMapp3rSegmentation.R
\name{hippMapp3rSegmentation}
\alias{hippMapp3rSegmentation}
\title{hippMapp3rSegmentation}
\usage{
hippMapp3rSegmentation(image, outputDirectory = NULL, verbose = FALSE)
}
\arguments{
\item{image}{input 3-D T1-weighted brain image.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
probability hippocampal mask (ANTsR image)
}
\description{
Perform HippMapp3r (hippocampal) segmentation described in
}
\details{
https://www.ncbi.nlm.nih.gov/pubmed/31609046

with models and architecture ported from

https://github.com/mgoubran/HippMapp3r

Additional documentation and attribution resources found at

https://hippmapp3r.readthedocs.io/en/latest/
}
\examples{
\dontrun{
library( ANTsRNet )
library( ANTsRNet )
library( keras )

url <- "https://github.com/mgoubran/HippMapp3r/blob/master/data/test_case/mprage.nii.gz?raw=true"
imageFile <- "head.nii.gz"
download.file( url, imageFile )
image <- antsImageRead( imageFile )
imageN4 <- n4BiasFieldCorrection( image, verbose = TRUE )
hippocampalProbabilityMask <- hippMapp3rSegmentation( imageN4, verbose = TRUE )
}
}
\author{
Tustison NJ
}
