% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{data}
\name{L2NormalizationLayer2D}
\alias{L2NormalizationLayer2D}
\title{L2 2-D normalization layer for SSD300/512 architecture.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
L2NormalizationLayer2D
}
\arguments{
\item{scale}{feature scale (default = 20).}
}
\value{
output tensor with the same shape as the input.
}
\description{
L2 2-D normalization layer for SSD300/512 architecture described in
}
\details{
Wei Liu, Andrew Rabinovich, and Alexander C. Berg.  ParseNet: Looking Wider
to See Better.

available here:

\code{https://arxiv.org/abs/1506.04579}
}
\author{
Tustison NJ
}
\keyword{datasets}
