% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\name{layer_l2_normalization_2d}
\alias{layer_l2_normalization_2d}
\title{Normalization layer (2-D)}
\usage{
layer_l2_normalization_2d(object, scale = 20, name = NULL,
  trainable = TRUE)
}
\arguments{
\item{scale}{box scale}
}
\value{
a keras layer tensor
}
\description{
Wraps a custom layer for the SSD network
}
