% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{layer_mixture_density}
\alias{layer_mixture_density}
\title{Mixture density layer}
\usage{
layer_mixture_density(
  objects,
  outputDimension,
  numberOfMixtures,
  trainable = TRUE
)
}
\arguments{
\item{outputDimension}{output dimension}

\item{numberOfMixtures}{number of Gaussians used to model the function}
}
\value{
a keras layer tensor
}
\description{
Wraps a custom mixture density layer.
}
