% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTensorUtilities.R
\name{layer_resample_tensor_2d}
\alias{layer_resample_tensor_2d}
\title{Creates a resampled tensor (to fixed size) layer (2-D)}
\usage{
layer_resample_tensor_2d(
  object,
  shape,
  interpolationType = c("nearestNeighbor", "nearest", "linear", "bilinear", "cubic",
    "bicubic"),
  name = NULL,
  trainable = FALSE
)
}
\arguments{
\item{object}{Object to compose layer with. This is either a
\link[keras:keras_model_sequential]{keras::keras_model_sequential} to add the layer to,
or another Layer which this layer will call.}

\item{shape}{vector or list of length 2 specifying the shape of the output tensor.}

\item{interpolationType}{type of interpolation for resampling.  Can be
\code{nearestNeighbor}, \code{nearest},
\code{linear}, \code{bilinear},
\code{cubic}, or \code{bicubic}.}

\item{name}{The name of the layer}

\item{trainable}{Whether the layer weights will be updated during training.}
}
\value{
a keras layer tensor
}
\description{
Resamples a spatial tensor based on the specified shape and interpolation type.
}
\author{
Tustison NJ
}
