% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesionSegmentation.R
\name{lesionSegmentation}
\alias{lesionSegmentation}
\title{wholeHeadInpainting}
\usage{
lesionSegmentation(t1, doPreprocessing = TRUE, verbose = FALSE)
}
\arguments{
\item{t1}{input 3-D T1 brain image (not skull-stripped).}

\item{doPreprocessing}{Perform n4 bias correction, intensity truncation, brain
extraction.}

\item{verbose}{print progress.}
}
\value{
inpainted image
}
\description{
Perform in-painting for whole-head MRI
}
\author{
Tustison NJ
}
