% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdUtilities.R
\docType{class}
\name{LossSSD}
\alias{LossSSD}
\title{Loss function for the SSD deep learning architecture.}
\value{
an SSD loss function
}
\description{
Creates an R6 class object for use with the SSD deep learning architecture
based on the paper
}
\details{
W. Liu, D. Anguelov, D. Erhan, C. Szegedy, S. Reed, C-Y. Fu, A. Berg.
SSD: Single Shot MultiBox Detector.

available here:

\url{https://arxiv.org/abs/1512.02325}
}
\section{Usage}{

\preformatted{ssdLoss <- LossSSD$new( dimension = 2L, backgroundRatio = 3L, 
  minNumberOfBackgroundBoxes = 0L,  alpha = 1.0, 
  numberOfClassificationLabels )

ssdLoss$smooth_l1_loss( y_true, y_pred )
ssdLoss$log_loss( y_true, y_pred )
ssdLoss$compute_loss( y_true, y_pred )
}
}

\section{Arguments}{

\describe{
\item{ssdLoss}{A \code{process} object.}
\item{dimension}{image dimensionality.}
\item{backgroundRatio}{The maximum ratio of background to foreround
for weighting in the loss function.  Is rounded to the nearest integer.
Default is 3.}
\item{minNumberOfBackgroundBoxes}{The minimum number of background boxes
to use in loss computation \emph{per batch}.  Should reflect a value in
proportion to the batch size.  Default is 0.}
\item{alpha}{Weighting factor for the localization loss in total loss
computation.}
\item{numberOfClassificationLabels}{number of classes including background.}
}
}

\section{Details}{

\code{$smooth_l1_loss} smooth loss

\code{$log_loss} log loss

\code{$compute_loss} computes total loss.
}

\author{
Tustison NJ
}
