% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lungExtraction.R
\name{lungExtraction}
\alias{lungExtraction}
\title{Lung extraction}
\usage{
lungExtraction(
  image,
  modality = c("proton", "ct"),
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D lung image.}

\item{modality}{image type.  Options include "proton" or "ct".}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
segmentation and probability images
}
\description{
Perform proton (H1) or CT lung extraction using a U-net architecture.
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "lung.nii.gz" )
output <- lungExtraction( image, modality = "proton" )
}
}
\author{
Tustison NJ
}
