% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{mixture_density_network_softmax}
\alias{mixture_density_network_softmax}
\title{Softmax function for mixture density with temperature adjustment}
\usage{
mixture_density_network_softmax(logits, temperature = 1)
}
\arguments{
\item{logits}{input of logits/mixture weights to adjust}

\item{temperature}{the temperature for to adjust the distribution (default 1.0)}
}
\value{
softmax loss value
}
\description{
Ported from:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}\if{html}{\out{</div>}}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
