% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.R
\name{mouseBrainParcellation}
\alias{mouseBrainParcellation}
\title{Mouse brain parcellation}
\usage{
mouseBrainParcellation(
  image,
  mask = NULL,
  returnIsotropicOutput = FALSE,
  whichParcellation = c("nick"),
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D brain image (or list of images for multi-modal scenarios).}

\item{mask}{Brain mask.  If not specified, one is estimated using ANTsXNet mouse
brain extraction.}

\item{returnIsotropicOutput}{The network actually learns an interpolating
function specific to the mouse brain.  Setting this to true, the output
images are returned isotropically resampled.}

\item{whichParcellation}{Brain parcellation type:
\itemize{
\item{"nick":
\itemize{
\item{Label 0:}{background}
\item{Label 1:}{CSF}
\item{Label 2:}{gray matter}
\item{Label 3:}{white matter}
\item{Label 4:}{deep gray matter}
\item{Label 5:}{brain stem}
\item{Label 6:}{cerebellum}
}}
}}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
brain probability mask
}
\description{
Perform brain extraction of mouse T2 MRI
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "brain.nii.gz" )
parcellation <- mouseBrainParcellation( image, modality = "t2" )
}
}
\author{
Tustison NJ
}
