% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mriSuperResolution.R
\name{mriSuperResolution}
\alias{mriSuperResolution}
\title{Super-resolution for MRI}
\usage{
mriSuperResolution(image, outputDirectory = NULL)
}
\arguments{
\item{image}{magnetic resonance image}
}
\value{
super-resolution image.
}
\description{
Perform super-resolution (2x) of MRI data using deep back projection network.
}
\examples{
\dontrun{
library( ANTsRNet )

image <- antsImageRead( "t1.nii.gz" )
imageSr <- mriSuperResolution( image )
}
}
\author{
Avants BB
}
