% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customLosses.R
\name{multilabel_dice_coefficient}
\alias{multilabel_dice_coefficient}
\title{Model loss function for multilabel problems--- multilabel dice coefficient}
\usage{
multilabel_dice_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\description{
Based on the keras loss function (losses.R):
}
\details{
\url{https://github.com/rstudio/keras/blob/master/R/losses.R}

Loss functions are to be supplied in the loss parameter of the
\code{compile()} function.

Loss functions can be specified either using the name of a built in loss
function (e.g. \code{loss = binary_crossentropy}), a reference to a built in loss
function (e.g. \code{loss = binary_crossentropy()}) or by passing an
arbitrary function that returns a scalar for each data-point.
The actual optimized objective is the mean of the output array across all
datapoints.
}
