% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{multilabel_dice_coefficient}
\alias{multilabel_dice_coefficient}
\title{Dice function for multilabel problems}
\usage{
multilabel_dice_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
Dice value
}
\description{
Dice function for multilabel problems
}
\examples{

library( ANTsR )
library( ANTsRNet )
library( keras )

domainImage <- antsImageRead( getANTsRData( "r16" ) )

# Create the model

model <- createUnetModel2D( c( dim( domainImage ), 1 ),
  numberOfOutputs = numberOfLabels )

metric_multilabel_dice_coefficient <-
  custom_metric( "multilabel_dice_coefficient",
    multilabel_dice_coefficient )

model \%>\% compile( loss = categorical_crossentropy,
  optimizer = optimizer_adam( lr = 0.0001 ),
  metrics = c( metric_multilabel_dice_coefficient ) )

}
\author{
Tustison NJ
}
