% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unetUtilities.R
\name{multilabel_dice_coefficient}
\alias{multilabel_dice_coefficient}
\title{Model loss function for multilabel problems--- multilabel dice coefficient

Taken the keras loss function (losses.R):}
\usage{
multilabel_dice_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\description{
https://github.com/rstudio/keras/blob/master/R/losses.R
}
\details{
Loss functions are to be supplied in the \code{loss} parameter of the
\code{\link[=compile]{compile()}} function.

Loss functions can be specified either using the name of a built in loss
function (e.g. 'loss = binary_crossentropy'), a reference to a built in loss
function (e.g. 'loss = loss_binary_crossentropy()') or by passing an
artitrary function that returns a scalar for each data-point and takes the
following two arguments:
\itemize{
\item \code{y_true} True labels (Tensor)
\item \code{y_pred} Predictions (Tensor of the same shape as \code{y_true})
}

The actual optimized objective is the mean of the output array across all
datapoints.
}
