% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{multilabel_dice_coefficient}
\alias{multilabel_dice_coefficient}
\title{Dice function for multilabel segmentation problems}
\usage{
multilabel_dice_coefficient(y_true, y_pred, smoothingFactor = 0)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}

\item{smoothingFactor}{parameter for smoothing the metric.}
}
\value{
Dice value (negative)
}
\description{
Dice function for multilabel segmentation problems
}
\examples{

library( ANTsRNet )
library( keras )

model <- createUnetModel2D( c( 64, 64, 1 ) )

dice_loss <- multilabel_dice_coefficient( smoothing_factor = 0.1 )

model \%>\% compile( loss = dice_loss,
  optimizer = optimizer_adam( lr = 0.0001 ) )
rm(model); gc()
}
\author{
Tustison NJ
}
