% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{peak_signal_to_noise_ratio}
\alias{peak_signal_to_noise_ratio}
\title{Model loss function for super-resolution---peak-signal-to-noise ratio.}
\usage{
peak_signal_to_noise_ratio(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\description{
Based on the keras loss function (losses.R):
}
\details{
\url{https://github.com/rstudio/keras/blob/master/R/losses.R}

Loss functions are to be supplied in the loss parameter of the
\code{compile()} function.
}
