% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customMetrics.R
\name{pearson_correlation_coefficient}
\alias{pearson_correlation_coefficient}
\title{Function for Pearson correlation coefficient.}
\usage{
pearson_correlation_coefficient(y_true, y_pred)
}
\arguments{
\item{y_true}{True labels (Tensor)}

\item{y_pred}{Predictions (Tensor of the same shape as \code{y_true})}
}
\value{
Correlation
}
\description{
Function for Pearson correlation coefficient.
}
\author{
Tustison NJ
}
