% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessImage.R
\name{preprocessBrainImage}
\alias{preprocessBrainImage}
\title{Basic preprocessing pipeline for T1-weighted brain MRI}
\usage{
preprocessBrainImage(
  image,
  truncateIntensity = c(0.01, 0.99),
  doBrainExtraction = TRUE,
  templateTransformType = NULL,
  template = "biobank",
  doBiasCorrection = TRUE,
  returnBiasField = FALSE,
  doDenoising = TRUE,
  intensityMatchingType = NULL,
  referenceImage = NULL,
  intensityNormalizationType = NULL,
  outputDirectory = NULL,
  verbose = TRUE
)
}
\arguments{
\item{image}{input t1-weighted brain MRI}

\item{truncateIntensity}{2-element vector giving the low and high quantiles
for intensity truncation.}

\item{doBrainExtraction}{boolean for performing ANTsRNet brain extraction
If \code{TRUE}, returns mask.  3-D input only.}

\item{templateTransformType}{see Details in help for \code{antsRegistration}.
Typically "Rigid" or "Affine".}

\item{template}{an ANTs image (not skull-stripped). Other premade templates
are "biobank" and "croppedMni152".}

\item{doBiasCorrection}{boolean for performing N4 bias field correction.}

\item{returnBiasField}{if TRUE, return bias field as an additional output
\emph{without} bias correcting the preprocessed image.}

\item{doDenoising}{boolean for performing non-local means denoising.}

\item{intensityMatchingType}{Either "regression" or "histogram".  Only is
performed if \code{!is.null(referenceImage)}.}

\item{referenceImage}{reference image for intensity matching.}

\item{intensityNormalizationType}{Either rescale the intensities to \link{0,1}
(i.e., "01") or zero-mean, unit variance (i.e., "0mean").  If \code{NULL}
normalization is not performed.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress to the screen.}
}
\value{
preprocessed image and, optionally, the brain mask, bias field, and
template transforms.
}
\description{
Various preprocessing steps that have been previously described in
various papers including the cortical thickness pipeline:
}
\details{
\preformatted{\url{https://www.ncbi.nlm.nih.gov/pubmed/24879923}
}
}
\examples{

library( ANTsR )
library( ANTsRNet )

image <- antsImageRead( getANTsRData( "r16" ) )
preprocessedImage <- preprocessBrainImage( image, 
  truncateIntensity = c( 0.01, 0.99 ), doBrainExtraction = FALSE,
  doBiasCorrection = TRUE, doDenoising = TRUE, 
  intensityNormalizationType = "01", verbose = FALSE )

}
\author{
Tustison NJ, Avants BB
}
