% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{sampleFromCategoricalDistribution}
\alias{sampleFromCategoricalDistribution}
\title{Sample from a categorical distribution}
\usage{
sampleFromCategoricalDistribution(distribution)
}
\arguments{
\item{distribution}{input categorical distribution from which
to sample.}
}
\value{
a single sample
}
\description{
Ported from:
}
\details{
\preformatted{    https://github.com/cpmpercussion/keras-mdn-layer/
}
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
