% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{sampleFromOutput}
\alias{sampleFromOutput}
\title{Sample from a distribution}
\usage{
sampleFromOutput(parameters, outputDimension, numberOfMixtures,
  temperature = 1, sigmaTemperature = 1)
}
\arguments{
\item{distribution}{input distribution from which to sample.}
}
\value{
a single sample
}
\description{
Ported from:
}
\details{
https://github.com/cpmpercussion/keras-mdn-layer/
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
