% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denseUnetUtilities.R
\docType{class}
\name{ScaleLayer}
\alias{ScaleLayer}
\title{Custom scale layer}
\value{
.
}
\description{
Custom scale layer
}
\section{Arguments}{

\describe{
\item{axis}{integer specifying which axis to normalize.}
\item{momentum}{momentum value used for computation of the exponential
average of the mean and standard deviation.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
