% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTransformerNetworkUtilities.R
\docType{class}
\name{SpatialTransformerLayer2D}
\alias{SpatialTransformerLayer2D}
\title{Spatial transformer layer (2-D)}
\value{
resampled batch images.
}
\description{
Spatial transformer layer (2-D)
}
\section{Arguments}{

\describe{
\item{inputs}{list of size 2 where the first element are the images. The second
element are the weights.}
\item{resampledSize}{size of the resampled output images.}
}
}

\section{Details}{

\code{$initialize} instantiates a new class.

\code{$call} main body.

\code{$compute_output_shape} computes the output shape.
}

\author{
Tustison NJ
}
