% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureDensityUtilities.R
\name{splitMixtureParameters}
\alias{splitMixtureParameters}
\title{Splits the mixture parameters.}
\usage{
splitMixtureParameters(parameters, outputDimension, numberOfMixes)
}
\arguments{
\item{parameters}{vector parameter to split}

\item{outputDimension}{output dimension}

\item{numberOfMixes}{number of mixture components}
}
\value{
separate mixture parameters
}
\description{
Ported from:
}
\details{
https://github.com/cpmpercussion/keras-mdn-layer/
}
\examples{

library( keras )


}
\author{
Tustison NJ
}
