% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superResolutionUtilities.R
\name{SSIM}
\alias{SSIM}
\title{Structural similarity index (SSI) between two images.}
\usage{
SSIM(x, y, K = c(0.01, 0.03))
}
\arguments{
\item{x}{input image.}

\item{y}{input image.}

\item{K}{vector of length 2 which contain SSI parameters meant to stabilize
the formula in case of weak denominators.}
}
\value{
the structural similarity index
}
\description{
Implementation of the SSI quantity for two images proposed in
}
\details{
Z. Wang, A.C. Bovik, H.R. Sheikh, E.P. Simoncelli. "Image quality
assessment: from error visibility to structural similarity". IEEE TIP.
13 (4): 600–612.
}
\examples{

library( ANTsR )

r16 <- antsImageRead( getANTsRData( 'r16' ) )
r85 <- antsImageRead( getANTsRData( 'r85' ) )
ssimValue <- SSIM( r16, r85 )

}
\author{
Avants BB
}
