% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteMatterHyperintensitySegmentation.R
\name{sysuMediaWmhSegmentation}
\alias{sysuMediaWmhSegmentation}
\title{White matter hyperintensity segmentation}
\usage{
sysuMediaWmhSegmentation(
  flair,
  t1 = NULL,
  doPreprocessing = TRUE,
  useEnsemble = TRUE,
  useAxialSlicesOnly = TRUE,
  outputDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{flair}{input 3-D FLAIR brain image.}

\item{t1}{input 3-D T1-weighted brain image (assumed to be aligned to
the flair, if specified).}

\item{doPreprocessing}{perform n4 bias correction?}

\item{useEnsemble}{boolean to check whether to use all 3 sets of weights.}

\item{useAxialSlicesOnly}{if \code{TRUE}, use original implementation which
was trained on axial slices.  If \code{FALSE}, use ANTsXNet variant
implementation which applies the slice-by-slice models to all 3 dimensions
and averages the results.}

\item{outputDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(outputDirectory)}, these data will be downloaded to the
inst/extdata/ subfolder of the ANTsRNet package.}

\item{verbose}{print progress.}
}
\value{
WMH segmentation probability image
}
\description{
Perform WMH segmentation using the winning submission in the MICCAI
2017 challenge by the sysu_media team using FLAIR or T1/FLAIR.  The
MICCAI challenge is discussed in
}
\details{
\url{https://pubmed.ncbi.nlm.nih.gov/30908194/}

with the sysu_media's team entry is discussed in

\url{https://pubmed.ncbi.nlm.nih.gov/30125711/}

with the original implementation available here:\preformatted{\url{https://github.com/hongweilibran/wmh_ibbmTum}
}
}
\examples{
\dontrun{
library( ANTsRNet )
library( keras )

image <- antsImageRead( "flair.nii.gz" )
probabilityMask <-sysuMediaWmhSegmentation( image )
}
}
\author{
Tustison NJ
}
