% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvaSeg.R
\name{uvaSeg}
\alias{uvaSeg}
\title{Unsupervised variational autoencoder segmentation}
\usage{
uvaSeg(image, model, k, mask)
}
\arguments{
\item{image}{input image}

\item{model}{the model output from \code{uvaSegTrain}}

\item{k}{number of clusters or cluster centers}

\item{mask}{defining output segmentation space}
}
\value{
segmentation and probability images are output
}
\description{
Trains a variational autoencoding with a convolutional network. This is
followed by k-means clustering to produce a segmentation and probabilities.
}
\examples{

library(ANTsR)
patch <- ri( 1 ) \%>\% resampleImage( 4 ) \%>\% iMath( "Normalize" ) \%>\%
  extractImagePatches( c(3,3), maxNumberOfPatches = 50 )
uvaSegModel = uvaSegTrain( patch, 6 )
tarImg = ri( 3 ) \%>\% resampleImage( 4 )
uvaSegmentation = uvaSeg(tarImg, uvaSegModel, k = 3, getMask( tarImg ) )

}
\author{
Avants BB
}
