% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvaSeg.R
\name{uvaSegTrain}
\alias{uvaSegTrain}
\title{Unsupervised variational autoencoder training}
\usage{
uvaSegTrain(patches, k, convControl)
}
\arguments{
\item{patches}{input patch list, see \code{extractImagePatches}}

\item{k}{number of embedding layers}

\item{convControl}{optional named list with control parameters ( see code )}
}
\value{
model is output
}
\description{
Trains a variational autoencoding with a convolutional network. This is
followed by k-means clustering to produce a segmentation and probabilities.
}
\examples{
library(ANTsR)
patch <- ri( 1 ) \%>\% resampleImage( 4 ) \%>\% iMath( "Normalize" ) \%>\%
  extractImagePatches( c(3,3), maxNumberOfPatches = 50 )
mdl = uvaSegTrain( patch, 3 )

}
\author{
Avants BB
}
