% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inpainting.R
\name{wholeHeadInpainting}
\alias{wholeHeadInpainting}
\title{wholeHeadInpainting}
\usage{
wholeHeadInpainting(
  image,
  roiMask,
  modality = "t1",
  mode = "axial",
  antsxnetCacheDirectory = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{input 3-D MR image.}

\item{roiMask}{binary mask image}

\item{modality}{Modality image type.  Options include: "t1": T1-weighted MRI,
"flair": FLAIR MRI.}

\item{mode}{Options include:  "sagittal": sagittal view network, "coronal":
coronal view network, "axial": axial view network, "average": average of all
canonical views, "meg": morphological erosion, greedy, iterative.}

\item{antsxnetCacheDirectory}{destination directory for storing the downloaded
template and model weights.  Since these can be resused, if
\code{is.null(antsxnetCacheDirectory)}, these data will be downloaded to the
subdirectory ~/.keras/ANTsXNet/.}

\item{verbose}{print progress.}
}
\value{
inpainted image
}
\description{
Perform in-painting for whole-head MRI
}
\author{
Tustison NJ
}
